Cholesky

JuLDL.make_transform_fac — Function

the first part of make transform matrix, usually in the shuo function to same time and computing resources

JuLDL.make_transform_matrix — Function

using cholesky decomposition to calculate transformation matrix from S to C or from C to S

JuLDL.make_transform_fac — Method

maketransformfac(::SparseMatrixCSC)

calculate first part of cholesky decomposition for sparse matrix

JuLDL.make_transform_fac — Method

maketransformfac(::Matrix)

calculate first part of cholesky decomposition for dense matrix

JuLDL.make_transform_matrix — Method

maketransformmatrix(::Union{LinearAlgebra.Cholesky, SuiteSparse.CHOLMOD.Factor}, ::Union{SparseMatrixCSC, Matrix}, ::Union{SparseMatrixCSC, Matrix})

second part of calculate cholesky decomposition transformation matrix

...

Arguments

output_format::Symbol=:auto: to force output format to dense(:dense) or sparse(:sparse), make it auto(:auto) to determined by the program

verbose::Bool=false: if verbose, more information will be printed out

Examples

  C = [1 1 1 1 0 0 0 0; 1 0 1 0 1 0 1 0; 0 0 0 0 1 1 1 1]
  S = [1 0 1 0; 1 1 0 0; 0 0 1 1]

  JuLDL.make_transform_matrix(C, S)

...

JuLDL.make_transform_matrix — Method

maketransformmatrix(::SparseMatrixCSC, ::Matrix)

using cholesky decomposition to calculate transformation matrix from X to Y, where X is a sparse matrix and Y is a dense matrix

...

Arguments

method::Symbol=:additive: shift mode whether :additive or others

shift::AbstractFloat=0.02: shift value

multiplier::AbstractFloat=1.01: multiplier value

output_format::Symbol=:auto: to force output format to dense(:dense) or sparse(:sparse), make it auto(:auto) to determined by the program

verbose::Bool=false: if verbose, more information will be printed out

Examples

  C = [1 1 1 1 0 0 0 0; 1 0 1 0 1 0 1 0; 0 0 0 0 1 1 1 1]
  S = [1 0 1 0; 1 1 0 0; 0 0 1 1]

  JuLDL.make_transform_matrix(C, S)

...

JuLDL.make_transform_matrix — Method

maketransformmatrix(::Matrix, ::Union{SparseMatrixCSC, Matrix})

using cholesky decomposition to calculate transformation matrix from X to Y, where X is a dense matrix and Y is either a dense matrix or a sparse matrix

...

Arguments

method::Symbol=:additive: shift mode whether :additive or others

shift::AbstractFloat=0.02: shift value

multiplier::AbstractFloat=1.01: multiplier value

output_format::Symbol=:auto: to force output format to dense(:dense) or sparse(:sparse), make it auto(:auto) to determined by the program

verbose::Bool=false: if verbose, more information will be printed out

Examples

  C = [1 1 1 1 0 0 0 0; 1 0 1 0 1 0 1 0; 0 0 0 0 1 1 1 1]
  S = [1 0 1 0; 1 1 0 0; 0 0 1 1]

  JuLDL.make_transform_matrix(C, S)

...

JuLDL.make_transform_matrix — Method

maketransformmatrix(::SparseMatrixCSC, ::SparseMatrixCSC)

using cholesky decomposition to calculate transformation matrix from X to Y, where X is a sparse matrix and Y is a sparse matrix

...

Arguments

method::Symbol=:additive: shift mode whether :additive or others

shift::AbstractFloat=0.02: shift value

multiplier::AbstractFloat=1.01: multiplier value

output_format::Symbol=:auto: to force output format to dense(:dense) or sparse(:sparse), make it auto(:auto) to determined by the program

verbose::Bool=false: if verbose, more information will be printed out

Examples

  C = [1 1 1 1 0 0 0 0; 1 0 1 0 1 0 1 0; 0 0 0 0 1 1 1 1]
  S = [1 0 1 0; 1 1 0 0; 0 0 1 1]

  JuLDL.make_transform_matrix(C, S)

...

JuLDL.format_matrix — Method

format_matrix(::Union{SparseMatrixCSC, Matrix}, ::Symbol, ::Bool)

convert ourput matrix to a dense matrix or sparse matrix