Evaluation

JudiLing.Comp_Acc_Struct — Type

A structure that stores information about comprehension accuracy.

JudiLing.eval_SC — Function

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Ideally the target words have highest correlations on the diagonal of the pertinent correlation matrices. Homophones support option is implemented.

JudiLing.eval_SC_loose — Function

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Count it as correct if one of the top k candidates is correct. Homophones support option is implemented.

JudiLing.accuracy_comprehension — Method

accuracy_comprehension(S, Shat, data)

Evaluate comprehension accuracy.

Obligatory Arguments

S::Matrix: the (gold standard) S matrix

Shat::Matrix: the (predicted) Shat matrix

data::DataFrame: the dataset

Optional Arguments

target_col::Union{String, Symbol}=:Words: the column name for target strings

base::Vector=nothing: base features (typically a lexeme)

inflections::Union{Nothing, Vector}=nothing: other features (typically in inflectional features)

Examples

accuracy_comprehension(
    S_train,
    Shat_train,
    latin_val,
    target_col=:Words,
    base=[:Lexeme],
    inflections=[:Person, :Number, :Tense, :Voice, :Mood]
    )

accuracy_comprehension(
    S_val,
    Shat_val,
    latin_train,
    target_col=:Words,
    base=["Lexeme"],
    inflections=[:Person, :Number, :Tense, :Voice, :Mood]
    )

JudiLing.eval_SC — Method

eval_SC(SChat::AbstractArray, SC::AbstractArray)

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Ideally the target words have highest correlations on the diagonal of the pertinent correlation matrices.

Obligatory Arguments

SChat::Union{SparseMatrixCSC, Matrix}: the Chat or Shat matrix

SC::Union{SparseMatrixCSC, Matrix}: the C or S matrix

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

R::Bool=false: if true, pairwise correlation matrix R is return

eval_SC(Chat_train, cue_obj_train.C)
eval_SC(Chat_val, cue_obj_val.C)
eval_SC(Shat_train, S_train)
eval_SC(Shat_val, S_val)

JudiLing.eval_SC — Method

eval_SC(SChat::AbstractArray, SC::AbstractArray, SC_rest::AbstractArray)

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Ideally the target words have highest correlations on the diagonal of the pertinent correlation matrices.

Note

The order is important. The fist gold standard matrix has to be corresponing to the SChat matrix, such as eval_SC(Shat_train, S_train, S_val) or eval_SC(Shat_val, S_val, S_train)

Obligatory Arguments

SChat::Union{SparseMatrixCSC, Matrix}: the Chat or Shat matrix

SC::Union{SparseMatrixCSC, Matrix}: the training/validation C or S matrix

SC_rest::Union{SparseMatrixCSC, Matrix}: the validation/training C or S matrix

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

R::Bool=false: if true, pairwise correlation matrix R is return

eval_SC(Chat_train, cue_obj_train.C, cue_obj_val.C)
eval_SC(Chat_val, cue_obj_val.C, cue_obj_train.C)
eval_SC(Shat_train, S_train, S_val)
eval_SC(Shat_val, S_val, S_train)

JudiLing.eval_SC — Method

eval_SC(SChat::AbstractArray, SC::AbstractArray, data::DataFrame, target_col::Union{String, Symbol})

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Ideally the target words have highest correlations on the diagonal of the pertinent correlation matrices. Support for homophones.

Obligatory Arguments

SChat::Union{SparseMatrixCSC, Matrix}: the Chat or Shat matrix

SC::Union{SparseMatrixCSC, Matrix}: the C or S matrix

data::DataFrame: datasets

target_col::Union{String, Symbol}: target column name

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

R::Bool=false: if true, pairwise correlation matrix R is return

eval_SC(Chat_train, cue_obj_train.C, latin, :Word)
eval_SC(Chat_val, cue_obj_val.C, latin, :Word)
eval_SC(Shat_train, S_train, latin, :Word)
eval_SC(Shat_val, S_val, latin, :Word)

JudiLing.eval_SC — Method

eval_SC(SChat::AbstractArray, SC::AbstractArray, SC_rest::AbstractArray, data::DataFrame, data_rest::DataFrame, target_col::Union{String, Symbol})

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Ideally the target words have highest correlations on the diagonal of the pertinent correlation matrices.

Note

The order is important. The fist gold standard matrix has to be corresponing to the SChat matrix, such as eval_SC(Shat_train, S_train, S_val, latin, :Word) or eval_SC(Shat_val, S_val, S_train, latin, :Word)

Obligatory Arguments

SChat::Union{SparseMatrixCSC, Matrix}: the Chat or Shat matrix

SC::Union{SparseMatrixCSC, Matrix}: the training/validation C or S matrix

SC_rest::Union{SparseMatrixCSC, Matrix}: the validation/training C or S matrix

data::DataFrame: the training/validation datasets

data_rest::DataFrame: the validation/training datasets

target_col::Union{String, Symbol}: target column name

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

R::Bool=false: if true, pairwise correlation matrix R is return

eval_SC(Chat_train, cue_obj_train.C, cue_obj_val.C, latin, :Word)
eval_SC(Chat_val, cue_obj_val.C, cue_obj_train.C, latin, :Word)
eval_SC(Shat_train, S_train, S_val, latin, :Word)
eval_SC(Shat_val, S_val, S_train, latin, :Word)

JudiLing.eval_SC — Method

eval_SC(SChat::AbstractArray, SC::AbstractArray, batch_size::Int64)

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Ideally the target words have highest correlations on the diagonal of the pertinent correlation matrices. For large datasets, pass batch_size to process evaluation in chucks.

Obligatory Arguments

SChat: the Chat or Shat matrix

SC: the C or S matrix

data: datasets

target_col: target column name

batch_size: batch size

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

verbose::Bool=false: if true, more information is printed

eval_SC(Chat_train, cue_obj_train.C, latin, :Word)
eval_SC(Chat_val, cue_obj_val.C, latin, :Word)
eval_SC(Shat_train, S_train, latin, :Word)
eval_SC(Shat_val, S_val, latin, :Word)

JudiLing.eval_SC — Method

eval_SC(SChat::AbstractArray, SC::AbstractArray, data::DataFrame, target_col::Union{String, Symbol}, batch_size::Int64)

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Ideally the target words have highest correlations on the diagonal of the pertinent correlation matrices. For large datasets, pass batch_size to process evaluation in chucks. Support homophones.

Obligatory Arguments

SChat::AbstractArray: the Chat or Shat matrix

SC::AbstractArray: the C or S matrix

data::DataFrame: datasets

target_col::Union{String, Symbol}: target column name

batch_size::Int64: batch size

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

verbose::Bool=false: if true, more information is printed

eval_SC(Chat_train, cue_obj_train.C, latin, :Word, 5000)
eval_SC(Chat_val, cue_obj_val.C, latin, :Word, 5000)
eval_SC(Shat_train, S_train, latin, :Word, 5000)
eval_SC(Shat_val, S_val, latin, :Word, 5000)

JudiLing.eval_SC_loose — Method

eval_SC_loose(SChat, SC, k)

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Count it as correct if one of the top k candidates is correct.

Obligatory Arguments

SChat::Union{SparseMatrixCSC, Matrix}: the Chat or Shat matrix

SC::Union{SparseMatrixCSC, Matrix}: the C or S matrix

k: top k candidates

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

eval_SC_loose(Chat, cue_obj.C, k)
eval_SC_loose(Shat, S, k)

JudiLing.eval_SC_loose — Method

eval_SC_loose(SChat, SC, k, data, target_col)

Assess model accuracy on the basis of the correlations of row vectors of Chat and C or Shat and S. Count it as correct if one of the top k candidates is correct. Support for homophones.

Obligatory Arguments

SChat::Union{SparseMatrixCSC, Matrix}: the Chat or Shat matrix

SC::Union{SparseMatrixCSC, Matrix}: the C or S matrix

k: top k candidates

data: datasets

target_col: target column name

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

eval_SC_loose(Chat, cue_obj.C, k, latin, :Word)
eval_SC_loose(Shat, S, k, latin, :Word)

JudiLing.eval_manual — Method

eval_manual(res, data, i2f)

Create extensive reports for the outputs from build_paths and learn_paths.

JudiLing.eval_acc — Method

eval_acc(res, gold_inds::Array)

Evaluate the accuracy of the results from learn_paths or build_paths.

Obligatory Arguments

res::Array: the results from learn_paths or build_paths

gold_inds::Array: the gold paths' indices

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

verbose::Bool=false: if true, more information is printed

Examples

# evaluation on training data
acc_train = JudiLing.eval_acc(
    res_train,
    cue_obj_train.gold_ind,
    verbose=false
)

# evaluation on validation data
acc_val = JudiLing.eval_acc(
    res_val,
    cue_obj_val.gold_ind,
    verbose=false
)

JudiLing.eval_acc — Method

eval_acc(res, cue_obj::Cue_Matrix_Struct)

Evaluate the accuracy of the results from learn_paths or build_paths.

Obligatory Arguments

res::Array: the results from learn_paths or build_paths

cue_obj::Cue_Matrix_Struct: the C matrix object

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

verbose::Bool=false: if true, more information is printed

Examples

acc = JudiLing.eval_acc(res, cue_obj)

JudiLing.eval_acc_loose — Method

eval_acc_loose(res, gold_inds)

Lenient evaluation of the accuracy of the results from learn_paths or build_paths, counting a prediction as correct when the correlation of the predicted and gold standard semantic vectors is among the n top correlations, where n is equal to max_can in the 'learnpaths' or `buildpaths` function.

Obligatory Arguments

res::Array: the results from learn_paths or build_paths

gold_inds::Array: the gold paths' indices

Optional Arguments

digits: the specified number of digits after the decimal place (or before if negative)

verbose::Bool=false: if true, more information is printed

Examples

# evaluation on training data
acc_train_loose = JudiLing.eval_acc_loose(
    res_train,
    cue_obj_train.gold_ind,
    verbose=false
)

# evaluation on validation data
acc_val_loose = JudiLing.eval_acc_loose(
    res_val,
    cue_obj_val.gold_ind,
    verbose=false
)

JudiLing.extract_gpi — Function

extract_gpi(gpi, threshold=0.1, tolerance=(-1000.0))

Extract, using gold paths' information, how many n-grams for a gold path are below the threshold but above the tolerance.