Make Cue Matrix

JudiLing.Cue_Matrix_Struct — Type

a struct that store info after makecuematrix C is the cue matrix f2i is the dictionary return indices giving features i2f is in another hand return features when giving indices gold_ind stores gold paths within a list of indices A is the adjacency matrix

JudiLing.make_cue_matrix — Function

Construct cue matrix.

JudiLing.make_ngrams — Function

Given tokens make n-grams.

JudiLing.make_cue_matrix — Method

makecuematrix(::DataFrame)

This function makes cue matrix and corresponding indices given dataset as csv file.

...

Arguments

grams::Integer=3: the number of grams for cues

target_col::Union{String, Symbol}=:Words: the column name for target

tokenized::Bool=false: whether the target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: what is the seperate token

keep_sep::Bool=false: whether to keep seperater in cues

start_end_token::Union{String, Char}="#": start and end token

verbose::Bool=false: if verbose, more information prints out

Examples

latin = CSV.DataFrame!(CSV.File(joinpath("data", "latin_mini.csv")))
latin_cue_obj_train = JudiLing.make_cue_matrix(
  latin,
  grams=3,
  target_col=:Word,
  tokenized=false,
  keep_sep=false
  )

...

JudiLing.make_cue_matrix — Method

makecuematrix(::DataFrame,::CueMatrixStruct)

This function make cue matrix and corresponding indices giving dataset as csv file and train dataset cue obj. This is often used to construct valcueobj, in order to maintain the same indices.

...

Arguments

grams::Integer=3: the number of grams for cues

target_col::Union{String, Symbol}=:Words: the column name for target

tokenized::Bool=false: whether the word is tokenized

sep_token::Union{Nothing, String, Char}=nothing: what is the seperate token

keep_sep::Bool=false: whether to keep seperater in cues

start_end_token::Union{String, Char}="#": start and end token

verbose::Bool=false: if verbose, more information prints out

Examples

latin = CSV.DataFrame!(CSV.File(joinpath("data", "latin_mini.csv")))
latin_cue_obj_train = JudiLing.make_cue_matrix(
  latin,
  grams=3,
  target_col=:Word,
  tokenized=false,
  keep_sep=false
  )
# simulate the val dataset. Notice here that latin_val is part of training dataset to make
# sure all features and n-grams covered by training dataset.
latin_val = latin[101:150,:]
latin_cue_obj_val = JudiLing.make_cue_matrix(
  latin_val,
  latin_cue_obj_train,
  grams=3,
  target_col=:Word,
  tokenized=false,
  keep_sep=false
  )

...

JudiLing.make_ngrams — Method

make_ngrams(::Array,::Integer,::Bool, ::Union{Nothing, String, Char},::Union{String, Char}

given a list of tokens, return all ngrams in a list