Make Semantic Matrix

JudiLing.PS_Matrix_Struct — Type

This a structure that store all information about prelinguistic and their feature indices. pS is the cue matrix f2i is the dictionary return indices giving features i2f is in another hand return features when giving indices

JudiLing.make_pS_matrix — Function

Make prelinguistic semantic matrix.

JudiLing.make_S_matrix — Function

Make simulated semantic matrix.

JudiLing.make_pS_matrix — Method

makepSmatrix(::DataFrame) -> ::PSMatrixStruct

Create prelinguistic matrix given a csv file.

...

Arguments

features_col::Symbol=:CommunicativeIntention: the column name for target

sep_token::String="_": the seperated token

Examples

s_obj_train = JudiLing.make_pS_matrix(
  utterance,
  features_col=:CommunicativeIntention,
  sep_token="_")

...

JudiLing.make_pS_matrix — Method

makepSmatrix(::DataFrame, PSMatrixStruct) -> ::PSMatrixStruct

Construct prelinguistic matrix giving utterances and training sobj. The feature indices should maintain the same as thoes in sobj.

...

Arguments

features_col::Symbol=:CommunicativeIntention: the column name for target

sep_token::String="_": the seperated token

Examples

s_obj_val = JudiLing.make_pS_matrix(
  utterance_val,
  s_obj_train,
  features_col=:CommunicativeIntention,
  sep_token="_")

...

JudiLing.make_S_matrix — Method

makeSmatrix(::DataFrame, ::Vector, ::Vector) -> ::Matrix

Create simulated semantic matrix. Give each feature a random semantic vector, and sum up all features to compose the semantic vector.

...

Arguments

ncol::Int64=200: the dimension size of vectors, usually the same as cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add noise at the end of construction

sd_noise::Int64=1: the sd of the noise matrix

normalized::Bool=false: if true, most of the values range between 1 and -1, it may exceeds 1 or -1 depending on the sd

Examples

# basic usage
S_train = JudiLing.make_S_matrix(
  french,
  ["Lexeme"],
  ["Tense","Aspect","Person","Number","Gender","Class","Mood"],
  ncol=200)

# deep mode
S_train = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  isdeep=true,
  ...)

# non-deep mode
S_train = JudiLing.make_S_matrix(
  ...
  isdeep=false,
  ...)

# add noise mode
S_train = JudiLing.make_S_matrix(
  ...
  add_noise=true,
  sd_noise=1,
  ...)

# controls of means and standard deviations
S_train = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  sd_base=4,
  sd_inflection=4,
  sd_noise=1,
  ...)

...

JudiLing.make_S_matrix — Method

makeSmatrix(::DataFrame, ::DataFrame, ::Vector, ::Vector) -> ::Tuple{Matrix, Matrix}

Create validation simulated semantic matrix. Give each feature a random semantic vector, and sum up all features to compose the semantic vector.

...

Arguments

ncol::Int64=200: the dimension size of vectors, usually the same as cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add noise at the end of construction

sd_noise::Int64=1: the sd of the noise matrix

normalized::Bool=false: if true, most of the values range between 1 and -1, it may exceeds 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_S_matrix(
  french,
  french_val,
  ["Lexeme"],
  ["Tense","Aspect","Person","Number","Gender","Class","Mood"],
  ncol=200)

# deep mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  isdeep=true,
  ...)

# non-deep mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  isdeep=false,
  ...)

# add noise mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  add_noise=true,
  sd_noise=1,
  ...)

# controls of means and standard deviations
S_train, S_val = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  sd_base=4,
  sd_inflection=4,
  sd_noise=1,
  ...)

...

JudiLing.make_S_matrix — Method

makeSmatrix(::DataFrame, ::Vector) -> ::Matrix

Create simulated semantic matrix, provided for dataset that only have base features. Give each feature a random semantic vector, and sum up all features to compose the semantic vector.

...

Arguments

ncol::Int64=200: the dimension size of vectors, usually the same as cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_base::Int64=4: the sd of base features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add noise at the end of construction

sd_noise::Int64=1: the sd of the noise matrix

normalized::Bool=false: if true, most of the values range between 1 and -1, it may exceeds 1 or -1 depending on the sd

Examples

# basic usage
S_train = JudiLing.make_S_matrix(
  french,
  ["Lexeme"],
  ncol=200)

# deep mode
S_train = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  isdeep=true,
  ...)

# non-deep mode
S_train = JudiLing.make_S_matrix(
  ...
  isdeep=false,
  ...)

# add noise mode
S_train = JudiLing.make_S_matrix(
  ...
  add_noise=true,
  sd_noise=1,
  ...)

# controls of means and standard deviations
S_train = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  sd_base=4,
  sd_inflection=4,
  sd_noise=1,
  ...)

...

JudiLing.make_S_matrix — Method

makeSmatrix(::DataFrame, ::DataFrame, ::Vector) -> ::Tuple{Matrix, Matrix}

Create validation simulated semantic matrix, provided for dataset that only have base features. Give each feature a random semantic vector, and sum up all features to compose the semantic vector.

...

Arguments

ncol::Int64=200: the dimension size of vectors, usually the same as cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_base::Int64=4: the sd of base features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add noise at the end of construction

sd_noise::Int64=1: the sd of the noise matrix

normalized::Bool=false: if true, most of the values range between 1 and -1, it may exceeds 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_S_matrix(
  french,
  french_val,
  ["Lexeme"],
  ncol=200)

# deep mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  isdeep=true,
  ...)

# non-deep mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  isdeep=false,
  ...)

# add noise mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  add_noise=true,
  sd_noise=1,
  ...)

# controls of means and standard deviations
S_train, S_val = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  sd_base=4,
  sd_inflection=4,
  sd_noise=1,
  ...)

...