Make Yt Matrix

JudiLing.make_Yt_matrix — Function

Make Yt matrix for timestep t.

JudiLing.make_Yt_matrix — Method

make_Yt_matrix(t, data, f2i)

Make Yt matrix for timestep t. A given column of the Yt matrix specifies the support for the corresponding n-gram predicted for timestep t for each of the observations (rows of Yt).

Obligatory Arguments

t::Int64: the timestep t

data::DataFrame: the dataset

f2i::Dict: the dictionary returning indices given features

Optional Arguments

tokenized::Bool=false: if true, the dataset target is assumed to be tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator token

verbose::Bool=false: if verbose, more information will be printed

Examples

latin = DataFrame(CSV.File(joinpath("data", "latin_mini.csv")))
JudiLing.make_Yt_matrix(2, latin)