Utils

JuLDL.iscorrect — Function

check whether the prediction is correct

JuLDL.display_pred — Function

display prediction nicely

JuLDL.translate — Function

translate indices into words or utterances

JuLDL.translate_path — Function

just append indices together

JuLDL.is_truly_sparse — Function

check whether a matrix is truly sparse regardless its format

check whether a matrix is truly sparse regardless its format

JuLDL.isattachable — Function

check a gram is attach to another gram

check a gram is attach to another gram

JuLDL.iscomplete — Function

check a gram could complete a path

JuLDL.isstart — Function

check a gram could start a path

JuLDL.isnovel — Function

check wheter a path is in training data or nor