Utils

JudiLing.wh_learn — Method

wh_learn(X, Y)

Widrow-Hoff Learning.

Obligatory Arguments

test_mode::Symbol: which test mode, currently supports :trainonly, :presplit, :carefulsplit and :randomsplit.

Optional Arguments

eta::Float64=0.1: the learning rate

n_epochs::Int64=1: the number of epochs to be trained

weights::Matrix=nothing: the initial weights

learn_seq::Vector=nothing: the learning sequence

save_history::Bool=false: if true, a partical training history will be saved

history_cols::Vector=nothing: the list of column indices you want to saved in history, e.g. [1,32,42] or [2]

history_rows::Vector=nothing: the list of row indices you want to saved in history, e.g. [1,32,42] or [2]

verbose::Bool = false: if true, more information will be printed out

JudiLing.make_learn_seq — Method

make_learn_seq(freq)

Make Widrow-Hoff learning sequence.