Cholesky

JudiLing.make_transform_fac — Function

The first part of make transform matrix, usually in the learn_paths function to same time and computing resources.

JudiLing.make_transform_matrix — Function

Using cholesky decomposition to calculate transformation matrix from S to C or from C to S.

JudiLing.make_transform_fac — Method

maketransformfac(::SparseMatrixCSC) -> ::SuiteSparse.CHOLMOD.Factor

Calculate first part of cholesky decomposition for sparse matrix.

JudiLing.make_transform_fac — Method

maketransformfac(::Matrix) -> ::LinearAlgebra.Cholesky

Calculate first part of cholesky decomposition for dense matrix.

JudiLing.make_transform_matrix — Method

maketransformmatrix(::Union{LinearAlgebra.Cholesky, SuiteSparse.CHOLMOD.Factor}, ::Union{SparseMatrixCSC, Matrix}, ::Union{SparseMatrixCSC, Matrix}) -> ::Union{Matrix, SparseMatrixCSC}

Second part of calculate cholesky decomposition transformation matrix.

...

Arguments

output_format::Symbol=:auto: to force output format to dense(:dense) or sparse(:sparse), make it auto(:auto) to determined by the program

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information will be printed out

...

JudiLing.make_transform_matrix — Method

maketransformmatrix(::SparseMatrixCSC, ::Matrix) -> ::Union{SparseMatrixCSC, Matrix}

Using cholesky decomposition to calculate transformation matrix from X to Y, where X is a sparse matrix and Y is a dense matrix.

...

Arguments

method::Symbol=:additive: shift mode whether :additive or :multiplicative

shift::AbstractFloat=0.02: shift value

multiplier::AbstractFloat=1.01: multiplier value

output_format::Symbol=:auto: to force output format to dense(:dense) or sparse(:sparse), make it auto(:auto) to determined by the program

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information will be printed out

Examples

JudiLing.make_transform_matrix(
  C,
  S,
  method=:additive,
  shift=0.02,
  output_format=:auto,
  sparse_ratio=0.2,
  verbose=false)

JudiLing.make_transform_matrix(
  C,
  S,
  method=:multiplicative,
  multiplier=1.01,
  output_format=:auto,
  sparse_ratio=0.2,
  verbose=false)

...

JudiLing.make_transform_matrix — Method

maketransformmatrix(::Matrix, ::Union{SparseMatrixCSC, Matrix}) -> ::Union{SparseMatrixCSC, Matrix}

Using cholesky decomposition to calculate transformation matrix from X to Y, where X is a dense matrix and Y is either a dense matrix or a sparse matrix.

...

Arguments

method::Symbol=:additive: shift mode whether :additive or :multiplicative

shift::AbstractFloat=0.02: shift value

multiplier::AbstractFloat=1.01: multiplier value

output_format::Symbol=:auto: to force output format to dense(:dense) or sparse(:sparse), make it auto(:auto) to determined by the program

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information will be printed out

Examples

JudiLing.make_transform_matrix(
  C,
  S,
  method=:additive,
  shift=0.02,
  output_format=:auto,
  sparse_ratio=0.2,
  verbose=false)

JudiLing.make_transform_matrix(
  C,
  S,
  method=:multiplicative,
  multiplier=1.01,
  output_format=:auto,
  sparse_ratio=0.2,
  verbose=false)

...

JudiLing.make_transform_matrix — Method

maketransformmatrix(::SparseMatrixCSC, ::SparseMatrixCSC) -> ::Union{SparseMatrixCSC, Matrix}

Using cholesky decomposition to calculate transformation matrix from X to Y, where X is a sparse matrix and Y is a sparse matrix.

...

Arguments

method::Symbol=:additive: shift mode whether :additive or :multiplicative

shift::AbstractFloat=0.02: shift value

multiplier::AbstractFloat=1.01: multiplier value

output_format::Symbol=:auto: to force output format to dense(:dense) or sparse(:sparse), make it auto(:auto) to determined by the program

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information will be printed out

Examples

JudiLing.make_transform_matrix(
  C,
  S,
  method=:additive,
  shift=0.02,
  output_format=:auto,
  sparse_ratio=0.2,
  verbose=false)

JudiLing.make_transform_matrix(
  C,
  S,
  method=:multiplicative,
  multiplier=1.01,
  output_format=:auto,
  sparse_ratio=0.2,
  verbose=false)

...

JudiLing.format_matrix — Method

format_matrix(::Union{SparseMatrixCSC, Matrix}, ::Symbol) -> ::Union{SparseMatrixCSC, Matrix}

Convert output matrix format to either a dense matrix or a sparse matrix.