Find Paths

JudiLing.eval_SC — Function

eval_SC(Union{SparseMatrixCSC, Matrix}, Union{SparseMatrixCSC, Matrix})

evaluate S and Shat or C and Chat

...

Examples

#after you had Shat and Chat
eval_SC(cue_obj_train.C, Chat_train)
eval_SC(cue_obj_val.C, Chat_val)
eval_SC(S_train, Shat_train)
eval_SC(S_val, Shat_val)

...

JudiLing.eval_acc — Method

eval_acc(::Array, ::Array)

Evaluate the outputs from shuo function or hua function

...

Arguments

verbose::Bool=false: if verbose, more information prints out

Examples

#after you had results from shuo or hua
acc_train = JudiLing.eval_acc(
  res_train,
  cue_obj_train.gold_ind,
  verbose=false
)
acc_val = JudiLing.eval_acc(
  res_val,
  cue_obj_val.gold_ind,
  verbose=false
)

...

JudiLing.eval_acc_loose — Method

evalaccloose(::Array, ::Array)

Evaluate the outputs from shuo function or hua function, if one of the candidates are correct, then we take it correct. This reflects how many paths we could found but we could not recogni

...

Arguments

verbose::Bool=false: if verbose, more information prints out

Examples

#after you had results from shuo or hua
acc_train_loose = JudiLing.eval_acc_loose(
  res_train,
  cue_obj_train.gold_ind,
  verbose=false
)
acc_val_loose = JudiLing.eval_acc_loose(
  res_val,
  cue_obj_val.gold_ind,
  verbose=false
)

...

JudiLing.eval_gpi — Function

Evaluate gold path info

JudiLing.eval_manual — Function

Manually check the results