Find Paths

JudiLing.Comp_Acc_Struct — Type

A structure that stores information about comprehension accuracy.

JudiLing.accuracy_comprehension — Function

accuracy_comprehension(::Matrix, ::Matrix) -> ::Comp_Acc_Struct

Evaluate the comprehension accuracy.

...

Obligatory Arguments

S::Matrix: the S matrix

Shat::Matrix: the Shat matrix

data::DataFrame: the dataset

Optional Arguments

target_col::Union{String, Symbol}=:Words: the column name for target strings

base::Vector=["Lexeme"]: base features

inflections::Union{Nothing, Vector}=nothing: inflectional features

Examples

accuracy_comprehension(
  S_train,
  Shat_train,
  latin_val,
  target_col=:Words,
  base=["Lexeme"],
  inflections=["Person","Number","Tense","Voice","Mood"]
  )

accuracy_comprehension(
  S_val,
  Shat_val,
  latin_train,
  target_col=:Words,
  base=["Lexeme"],
  inflections=["Person","Number","Tense","Voice","Mood"]
  )

...

JudiLing.eval_SC — Function

eval_SC(Union{SparseMatrixCSC, Matrix}, Union{SparseMatrixCSC, Matrix}) -> ::Float64

Quickly evaluate Shat and Chat matrices.

...

Obligatory Arguments

SChat::Union{SparseMatrixCSC, Matrix}: the Chat or Shat matrix

SC::Union{SparseMatrixCSC, Matrix}: the C or S matrix

Optional Arguments

#after you had Shat and Chat
eval_SC(cue_obj_train.C, Chat_train)
eval_SC(cue_obj_val.C, Chat_val)
eval_SC(S_train, Shat_train)
eval_SC(S_val, Shat_val)

...

JudiLing.eval_acc — Method

eval_acc(::Array, ::Array) -> ::Float64

Evaluate the accuracy of the results from learn_paths or build_paths.

...

Obligatory Arguments

res::Array: the results from learn_paths or build_paths

gold_inds::Array: the gold paths' indices

Optional Arguments

verbose::Bool=false: if true, more information is printed

Examples

#after you had results from learn_paths or build_paths
acc_train = JudiLing.eval_acc(
  res_train,
  cue_obj_train.gold_ind,
  verbose=false
)
acc_val = JudiLing.eval_acc(
  res_val,
  cue_obj_val.gold_ind,
  verbose=false
)

...

JudiLing.eval_acc_loose — Method

eval_acc_loose(::Array, ::Array) -> ::Float64

Evaluate the accuracy of the results from learn_paths or build_paths, if one of the candidates is correct, then we take it as correct. This reflects how many paths were found but could not be recognized as the best path.

...

Obligatory Arguments

res::Array: the results from learn_paths or build_paths

gold_inds::Array: the gold paths' indices

Optional Arguments

verbose::Bool=false: if true, more information is printed

Examples

#after you had results from learn_paths or build_paths
acc_train_loose = JudiLing.eval_acc_loose(
  res_train,
  cue_obj_train.gold_ind,
  verbose=false
)
acc_val_loose = JudiLing.eval_acc_loose(
  res_val,
  cue_obj_val.gold_ind,
  verbose=false
)

...

JudiLing.eval_gpi — Function

eval_gpi(::Vector{Gold_Path_Info_Struct}, ::Float64, ::Float64) -> ::Array

Evaluate gold paths' information. It summarizes how many n-grams for a gold path is below threshold but above tolerance.

JudiLing.eval_manual — Function

eval_manual(::Array, ::DataFrame, ::Dict) -> ::Nothing

Evaluate the results from build_paths and learn_paths manually.