Find Paths

JudiLing.Result_Path_Info_Struct — Type

store paths information found by learnpaths or buildpaths function

JudiLing.Gold_Path_Info_Struct — Type

store gold paths information including indices and indices support and total support it can be used to evaluate how low the threshold is set in order to find the correct paths

JudiLing.learn_paths — Function

learnpaths function takes each timestep individually and calculate Ythat respectively,

JudiLing.build_paths — Function

build_paths function is shortcut algorithms that only takes n-grams that closed to the validation data

JudiLing.learn_paths — Method

learn_paths(::DataFrame,::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict)

learnpaths function takes each timestep individually and calculate Ythat respectively,

...

Arguments

gold_ind::Union{Nothing, Vector}=nothing: for in goldpathinfo mode

Shat_val::Union{Nothing, Matrix}=nothing: for goldpathinfo mode

check_gold_path::Bool=false: if turn on goldpathinfo mode

max_t::Integer=15: maximum timestep

max_can::Integer=10: maximum candidates when output

threshold::AbstractFloat=0.1: for each timestep, only grams greater than threshold will be selected

is_tolerant::Bool=false: if in tolerant mode, path allows limited nodes under threshold but greater than tolerance

tolerance::AbstractFloat=(-1000.0): in tolerant mode, only nodes greater than tolerance and lesser than threshold will be selected

max_tolerance::Integer=4: maximum numbers of nodes allowed in a path

grams::Integer=3: n-grams

tokenized::Bool=false: whether tokenized

sep_token::Union{Nothing, String, Char}=nothing: seperate token

keep_sep::Bool=false: whether keep seperaters in grams

target_col::Union{String, :Symbol}=:Words: word column names

issparse::Symbol=:auto: mt matrix output format mode

verbose::Bool=false: if verbose, more information will be printed out

Examples

latin_train = CSV.DataFrame!(CSV.File(joinpath("data", "latin_mini.csv")))
cue_obj_train = JudiLing.make_cue_matrix(
  latin_train,
  grams=3,
  target_col=:Word,
  tokenized=false,
  keep_sep=false
  )

latin_val = latin_train[101:150,:]
cue_obj_val = JudiLing.make_cue_matrix(
  latin_val,
  cue_obj_train,
  grams=3,
  target_col=:Word,
  tokenized=false,
  keep_sep=false
  )

n_features = size(cue_obj_train.C, 2)

S_train, S_val = JudiLing.make_S_matrix(
  latin_train,
  latin_val,
  ["Lexeme"],
  ["Person","Number","Tense","Voice","Mood"],
  ncol=n_features)

G_train = JudiLing.make_transform_matrix(S_train, cue_obj_train.C)

Chat_train = S_train * G_train
Chat_val = S_val * G_train

F_train = JudiLing.make_transform_matrix(cue_obj_train.C, S_train)

Shat_train = cue_obj_train.C * F_train
Shat_val = cue_obj_val.C * F_train

A = cue_obj_train.A

max_t = JudiLing.cal_max_timestep(latin_train, latin_val, :Word)

res_train, gpi_train = JudiLing.learn_paths(
  latin_train,
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  gold_ind=cue_obj_train.gold_ind,
  Shat_val=Shat_train,
  check_gold_path=true,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  sep_token="_",
  keep_sep=false,
  target_col=:Word,
  issparse=:dense,
  verbose=false)

res_val, gpi_val = JudiLing.learn_paths(
  latin_train,
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  gold_ind=cue_obj_val.gold_ind,
  Shat_val=Shat_val,
  check_gold_path=true,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  is_tolerant=true,
  tolerance=0.1,
  max_tolerance=0,
  tokenized=false,
  sep_token="-",
  keep_sep=false,
  target_col=:Word,
  issparse=:dense,
  verbose=false)

...

JudiLing.build_paths — Method

build_paths(::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict,::Array)

build_paths function is shortcut algorithms that only takes n-grams that closed to the validation data

...

Arguments

rC::Union{Nothing, Matrix}=nothing: correlation Matrix of C and Chat, passing it to save computing time

max_t::Integer=15: maximum timestep

max_can::Integer=10: maximum candidates when output

n_neighbors::Integer=10: find indices only in top n neighbors

grams::Integer=3: n-grams

tokenized::Bool=false: whether tokenized

sep_token::Union{Nothing, String, Char}=nothing: seperate token

target_col::Union{String, :Symbol}=:Words: word column names

verbose::Bool=false: if verbose, more information will be printed out

Examples

latin_train = CSV.DataFrame!(CSV.File(joinpath("data", "latin_mini.csv")))
cue_obj_train = JudiLing.make_cue_matrix(
  latin_train,
  grams=3,
  target_col=:Word,
  tokenized=false,
  keep_sep=false
  )

latin_val = latin_train[101:150,:]
cue_obj_val = JudiLing.make_cue_matrix(
  latin_val,
  cue_obj_train,
  grams=3,
  target_col=:Word,
  tokenized=false,
  keep_sep=false
  )

n_features = size(cue_obj_train.C, 2)

S_train, S_val = JudiLing.make_S_matrix(
  latin_train,
  latin_val,
  ["Lexeme"],
  ["Person","Number","Tense","Voice","Mood"],
  ncol=n_features)

G_train = JudiLing.make_transform_matrix(S_train, cue_obj_train.C)

Chat_train = S_train * G_train
Chat_val = S_val * G_train

F_train = JudiLing.make_transform_matrix(cue_obj_train.C, S_train)

Shat_train = cue_obj_train.C * F_train
Shat_val = cue_obj_val.C * F_train

A = cue_obj_train.A

max_t = JudiLing.cal_max_timestep(latin_train, latin_val, :Word)

JudiLing.build_paths(
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

JudiLing.build_paths(
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

...

JudiLing.eval_can — Method

eval_can(::Vector{Vector{Tuple{Vector{Integer}, Integer}}},::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Dict,::Integer,::Bool)

at the end of finding path algorithms, each candidates need to be evaluated regarding their predicted semantic vectors

JudiLing.find_top_feature_indices — Method

findtopfeature_indices(::Matrix, ::Array)

find out all indices within the closed top n datarow for a given validation datarow