Find Paths

JudiLing.Result_Path_Info_Struct — Type

Store paths' information built by learn_paths or build_paths

JudiLing.Gold_Path_Info_Struct — Type

Store gold paths' information including indices and indices' support and total support. It can be used to evaluate how low the threshold needs to be set in order to find most of the correct paths or if set very low, all of the correct paths.

JudiLing.Threshold_Stat_Struct — Type

Store threshold and tolerance proportional for each timestep.

JudiLing.learn_paths — Function

A sequence finding algorithm using discrimination learning to predict, for a given word, which n-grams are best supported for a given position in the sequence of n-grams.

JudiLing.build_paths — Function

The build_paths function constructs paths by only considering those n-grams that are close to the target. It first takes the predicted c-hat vector and finds the closest n neighbors in the C matrix. Then it selects all n-grams of these neighbors, and constructs all valid paths with those n-grams. The path producing the best correlation with the target semantic vector (through synthesis by analysis) is selected.

Missing docstring.

Missing docstring for learn_paths(data, cue_obj, S_val, F_train, Chat_val). Check Documenter's build log for details.

Missing docstring.

Missing docstring for learn_paths(data_train, data_val, C_train, S_val, F_train, Chat_val, A, i2f, f2i). Check Documenter's build log for details.

JudiLing.build_paths — Method

build_paths(data_val, C_train, S_val, F_train, Chat_val, A, i2f, C_train_ind)

The build_paths function constructs paths by only considering those n-grams that are close to the target. It first takes the predicted c-hat vector and finds the closest n neighbors in the C matrix. Then it selects all n-grams of these neighbors, and constructs all valid paths with those n-grams. The path producing the best correlation with the target semantic vector (through synthesis by analysis) is selected.

Obligatory Arguments

data::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

C_train::SparseMatrixCSC: the C matrix for the training dataset

S_val::Union{SparseMatrixCSC, Matrix}: the S matrix for the validation dataset

F_train::Union{SparseMatrixCSC, Matrix}: the F matrix for the training dataset

Chat_val::Matrix: the Chat matrix for the validation dataset

A::SparseMatrixCSC: the adjacency matrix

i2f::Dict: the dictionary returning features given indices

C_train_ind::Array: the gold paths' indices for the training dataset

Optional Arguments

rC::Union{Nothing, Matrix}=nothing: correlation Matrix of C and Chat, specify to save computing time

max_t::Int64=15: maximum number of timesteps

max_can::Int64=10: maximum number of candidates to consider

n_neighbors::Int64=10: the top n form neighbors to be considered

grams::Int64=3: the number n of grams that make up n-grams

tokenized::Bool=false: if true, the dataset target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator

target_col::Union{String, :Symbol}=:Words: the column name for target strings

if_pca::Bool=false: turn on to enable pca mode

pca_eval_M::Matrix=nothing: pass original F for pca mode

verbose::Bool=false: if true, more information will be printed

Examples

# training dataset
JudiLing.build_paths(
    latin_train,
    cue_obj_train.C,
    S_train,
    F_train,
    Chat_train,
    A,
    cue_obj_train.i2f,
    cue_obj_train.gold_ind,
    max_t=max_t,
    n_neighbors=10,
    verbose=false
    )

# validation dataset
JudiLing.build_paths(
    latin_val,
    cue_obj_train.C,
    S_val,
    F_train,
    Chat_val,
    A,
    cue_obj_train.i2f,
    cue_obj_train.gold_ind,
    max_t=max_t,
    n_neighbors=10,
    verbose=false
    )

# pca mode
res_build = JudiLing.build_paths(
    korean,
    Array(Cpcat),
    S,
    F,
    ChatPCA,
    A,
    cue_obj.i2f,
    cue_obj.gold_ind,
    max_t=max_t,
    if_pca=true,
    pca_eval_M=Fo,
    n_neighbors=3,
    verbose=true
    )

Missing docstring.

Missing docstring for learn_paths_rpi(data_train, data_val, C_train, S_val, F_train, Chat_val, A, i2f, f2i). Check Documenter's build log for details.

JudiLing.eval_can — Method

eval_can(candidates, S, F::Union{Matrix,SparseMatrixCSC, Chain}, i2f, max_can, if_pca, pca_eval_M)

Calculate for each candidate path the correlation between predicted semantic vector and the gold standard semantic vector, and select as target for production the path with the highest correlation.

JudiLing.find_top_feature_indices — Method

find_top_feature_indices(rC, C_train_ind)

Find all indices for the n-grams of the top n closest neighbors of a given target.

JudiLing.make_ngrams_ind — Method

make_ngrams_ind(res, n)

Construct ngrams indices.