Find Paths

JudiLing.Result_Path_Info_Struct — Type

Store path information build by learnpaths() or buildpaths()

JudiLing.Gold_Path_Info_Struct — Type

Store gold path information including indices and indices' support and total support. It can be used to evaluate how low the threshold needs to be set in order to find the correct paths.

JudiLing.learn_paths — Function

learn_paths(::DataFrame, ::DataFrame, ::SparseMatrixCSC, ::Union{SparseMatrixCSC, Matrix}, ::Union{SparseMatrixCSC, Matrix}, ::Matrix, ::SparseMatrixCSC, ::Dict) -> ::Union{Tuple{Vector{Vector{Result_Path_Info_Struct}}, Vector{Gold_Path_Info_Struct}}, Vector{Vector{Result_Path_Info_Struct}}}

One sequence finding algorithm used discrimination learning for the position of triphones.

...

Obligatory Arguments

data::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

C_train::SparseMatrixCSC: the C matrix for training dataset

S_val::Union{SparseMatrixCSC, Matrix}: the S matrix for validation dataset

F_train::Union{SparseMatrixCSC, Matrix}: the F matrix for training dataset

Chat_val::Matrix: the Chat matrix for validation dataset

A::SparseMatrixCSC: the adjacency matrix

i2f::Dict: the dictionary return features given indices

Optional Arguments

gold_ind::Union{Nothing, Vector}=nothing: gold paths' indices

Shat_val::Union{Nothing, Matrix}=nothing: the Shat matrix for validation dataset

check_gold_path::Bool=false: if true, return a list of gold paths' information as second output

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum candidates to keep in the results

threshold::Float64=0.1:the value set for the support such that if the support of a n-gram is higher than this value, select the n-gram anyway

is_tolerant::Bool=false: if true, select a specified number of n-grams whose supports are below threshold and above tolerance to be added to the path

tolerance::Float64=(-1000.0): the value set in tolerant mode such that if the support for a n-gram is inbetween this value and the threshold and the max_tolerance number has not been reached, then allow this n-gram to be added to the path

max_tolerance::Int64=4: maximum number of nodes allowed in a path

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: if true, the dataset target is assumed to be tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator

keep_sep::Bool=false:if true, keep separators in cues

target_col::Union{String, :Symbol}=:Words: the column name for target strings

issparse::Symbol=:auto: control of whether output of Mt matrix is a dense matrix or a sparse matrix

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information is printed

Examples

# basic usage with non-tokenized data
res = JudiLing.learn_paths(
  latin,
  latin,
  cue_obj.C,
  S,
  F,
  Chat,
  A,
  cue_obj.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  keep_sep=false,
  target_col=:Word,
  verbose=true)

# basic usage with tokenized data
res = JudiLing.learn_paths(
  french,
  french,
  cue_obj.C,
  S,
  F,
  Chat,
  A,
  cue_obj.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=true,
  sep_token="-",
  keep_sep=true,
  target_col=:Syllables,
  verbose=true)

# basic usage for val data
res_val = JudiLing.learn_paths(
  latin_train,
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  keep_sep=false,
  target_col=:Word,
  verbose=true)

# turn on tolerance mode
res_val = JudiLing.learn_paths(
  ...
  threshold=0.1,
  is_tolerant=true,
  tolerance=-0.1,
  max_tolerance=4,
  ...)

# control over Sparsity
res_val = JudiLing.learn_paths(
  ...
  issparse=:auto,
  sparse_ratio=0.2,
  ...)

...

JudiLing.build_paths — Function

build_paths(::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict,::Array) -> ::Vector{Vector{Result_Path_Info_Struct}}

build_paths function is a shortcut algorithm for finding paths that only takes the n-grams of that which are close to the targets.

...

Obligatory Arguments

data::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

C_train::SparseMatrixCSC: the C matrix for training dataset

S_val::Union{SparseMatrixCSC, Matrix}: the S matrix for validation dataset

F_train::Union{SparseMatrixCSC, Matrix}: the F matrix for training dataset

Chat_val::Matrix: the Chat matrix for validation dataset

A::SparseMatrixCSC: the adjacency matrix

i2f::Dict: the dictionary return features given indices

C_train_ind::Array: the gold paths' indices for training dataset

Optional Arguments

rC::Union{Nothing, Matrix}=nothing: correlation Matrix of C and Chat, pass it to save computing time

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum candidates to keep in the results

n_neighbors::Int64=10: find indices only in top n neighbors

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: if true, the dataset target is assumed to be tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator

target_col::Union{String, :Symbol}=:Words: the column name for target strings

verbose::Bool=false: if true, more information is printed

Examples

# training dataset
JudiLing.build_paths(
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

# validation dataset
JudiLing.build_paths(
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

...

JudiLing.learn_paths — Method

learn_paths(::DataFrame, ::DataFrame, ::SparseMatrixCSC, ::Union{SparseMatrixCSC, Matrix}, ::Union{SparseMatrixCSC, Matrix}, ::Matrix, ::SparseMatrixCSC, ::Dict) -> ::Union{Tuple{Vector{Vector{Result_Path_Info_Struct}}, Vector{Gold_Path_Info_Struct}}, Vector{Vector{Result_Path_Info_Struct}}}

One sequence finding algorithm used discrimination learning for the position of triphones.

...

Obligatory Arguments

data::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

C_train::SparseMatrixCSC: the C matrix for training dataset

S_val::Union{SparseMatrixCSC, Matrix}: the S matrix for validation dataset

F_train::Union{SparseMatrixCSC, Matrix}: the F matrix for training dataset

Chat_val::Matrix: the Chat matrix for validation dataset

A::SparseMatrixCSC: the adjacency matrix

i2f::Dict: the dictionary return features given indices

Optional Arguments

gold_ind::Union{Nothing, Vector}=nothing: gold paths' indices

Shat_val::Union{Nothing, Matrix}=nothing: the Shat matrix for validation dataset

check_gold_path::Bool=false: if true, return a list of gold paths' information as second output

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum candidates to keep in the results

threshold::Float64=0.1:the value set for the support such that if the support of a n-gram is higher than this value, select the n-gram anyway

is_tolerant::Bool=false: if true, select a specified number of n-grams whose supports are below threshold and above tolerance to be added to the path

tolerance::Float64=(-1000.0): the value set in tolerant mode such that if the support for a n-gram is inbetween this value and the threshold and the max_tolerance number has not been reached, then allow this n-gram to be added to the path

max_tolerance::Int64=4: maximum number of nodes allowed in a path

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: if true, the dataset target is assumed to be tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator

keep_sep::Bool=false:if true, keep separators in cues

target_col::Union{String, :Symbol}=:Words: the column name for target strings

issparse::Symbol=:auto: control of whether output of Mt matrix is a dense matrix or a sparse matrix

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information is printed

Examples

# basic usage with non-tokenized data
res = JudiLing.learn_paths(
  latin,
  latin,
  cue_obj.C,
  S,
  F,
  Chat,
  A,
  cue_obj.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  keep_sep=false,
  target_col=:Word,
  verbose=true)

# basic usage with tokenized data
res = JudiLing.learn_paths(
  french,
  french,
  cue_obj.C,
  S,
  F,
  Chat,
  A,
  cue_obj.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=true,
  sep_token="-",
  keep_sep=true,
  target_col=:Syllables,
  verbose=true)

# basic usage for val data
res_val = JudiLing.learn_paths(
  latin_train,
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  keep_sep=false,
  target_col=:Word,
  verbose=true)

# turn on tolerance mode
res_val = JudiLing.learn_paths(
  ...
  threshold=0.1,
  is_tolerant=true,
  tolerance=-0.1,
  max_tolerance=4,
  ...)

# control over Sparsity
res_val = JudiLing.learn_paths(
  ...
  issparse=:auto,
  sparse_ratio=0.2,
  ...)

...

JudiLing.build_paths — Method

build_paths(::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict,::Array) -> ::Vector{Vector{Result_Path_Info_Struct}}

build_paths function is a shortcut algorithm for finding paths that only takes the n-grams of that which are close to the targets.

...

Obligatory Arguments

data::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

C_train::SparseMatrixCSC: the C matrix for training dataset

S_val::Union{SparseMatrixCSC, Matrix}: the S matrix for validation dataset

F_train::Union{SparseMatrixCSC, Matrix}: the F matrix for training dataset

Chat_val::Matrix: the Chat matrix for validation dataset

A::SparseMatrixCSC: the adjacency matrix

i2f::Dict: the dictionary return features given indices

C_train_ind::Array: the gold paths' indices for training dataset

Optional Arguments

rC::Union{Nothing, Matrix}=nothing: correlation Matrix of C and Chat, pass it to save computing time

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum candidates to keep in the results

n_neighbors::Int64=10: find indices only in top n neighbors

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: if true, the dataset target is assumed to be tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator

target_col::Union{String, :Symbol}=:Words: the column name for target strings

verbose::Bool=false: if true, more information is printed

Examples

# training dataset
JudiLing.build_paths(
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

# validation dataset
JudiLing.build_paths(
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

...

JudiLing.eval_can — Method

eval_can(::Vector{Vector{Tuple{Vector{Int64}, Int64}}},::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Dict,::Int64,::Bool) -> ::Array{Array{Result_Path_Info_Struct,1},1}

Evaluate each candidate regarding the correlation between predicted semantic vector and original semantic vector.

JudiLing.find_top_feature_indices — Method

find_top_feature_indices(::Matrix, ::Array) -> ::Vector{Vector{Int64}}

Find all indices of the top n closest neighbors for a given target.