Find Paths

JudiLing.Result_Path_Info_Struct — Type

Store path information found by learnpaths() or buildpaths()

JudiLing.Gold_Path_Info_Struct — Type

Store gold path information including indices and indices' support and total support. It can be used to evaluate how low the threshold needs to be set in order to find the correct paths.

JudiLing.learn_paths — Function

learnpaths(::DataFrame,::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict) -> ::::Union{Tuple{Vector{Vector{ResultPathInfoStruct}}, Vector{GoldPathInfoStruct}}, Vector{Vector{ResultPathInfoStruct}}}

Take individual timestep and calculate its corresponding Yt_hat.

...

Arguments

gold_ind::Union{Nothing, Vector}=nothing: for in goldpathinfo mode

Shat_val::Union{Nothing, Matrix}=nothing: for goldpathinfo mode

check_gold_path::Bool=false: if true, turn on goldpathinfo mode

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum candidates to keep in the results

threshold::AbstractFloat=0.1:the value set for the support such that if the support of a n-gram is higher than this value, select the n-gram anyway

is_tolerant::Bool=false: if true, select a specified number of n-grams whose supports are below threshold and above tolerance to be added to the path

tolerance::AbstractFloat=(-1000.0): the value set in tolerant mode such that if the support for a n-gram is inbetween this value and the threshold and the max_tolerance number has not been reached, then allow this n-gram to be added to the path

max_tolerance::Int64=4: maximum number of nodes allowed in a path

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: if true, the dataset target is assumed to be tokenized

sep_token::Union{Nothing, String, Char}=nothing: separate token

keep_sep::Bool=false:if true, keep separators in results as well

target_col::Union{String, :Symbol}=:Words: target column names

issparse::Symbol=:auto: Mt matrix output format mode

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information will be printed out

Examples

res_train, gpi_train = JudiLing.learn_paths(
  latin_train,
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  gold_ind=cue_obj_train.gold_ind,
  Shat_val=Shat_train,
  check_gold_path=true,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  sep_token="_",
  keep_sep=false,
  target_col=:Word,
  issparse=:dense,
  verbose=false)

res_val, gpi_val = JudiLing.learn_paths(
  latin_train,
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  gold_ind=cue_obj_val.gold_ind,
  Shat_val=Shat_val,
  check_gold_path=true,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  is_tolerant=true,
  tolerance=0.1,
  max_tolerance=0,
  tokenized=false,
  sep_token="-",
  keep_sep=false,
  target_col=:Word,
  issparse=:dense,
  verbose=false)

...

JudiLing.build_paths — Function

buildpaths(::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict,::Array) -> ::Vector{Vector{ResultPathInfoStruct}}

build_paths function is a shortcut algorithm for finding paths that only takes the n-grams of that which are close to the targets.

...

Arguments

rC::Union{Nothing, Matrix}=nothing: correlation Matrix of C and Chat, pass it to save computing time

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum candidates to keep in the results

n_neighbors::Int64=10: find indices only in top n neighbors

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: if true, the dataset target is assumed to be tokenized

sep_token::Union{Nothing, String, Char}=nothing: separate token

target_col::Union{String, :Symbol}=:Words: target column names

verbose::Bool=false: if true, more information will be printed out

Examples

JudiLing.build_paths(
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

JudiLing.build_paths(
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

...

JudiLing.learn_paths — Method

learnpaths(::DataFrame,::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict) -> ::::Union{Tuple{Vector{Vector{ResultPathInfoStruct}}, Vector{GoldPathInfoStruct}}, Vector{Vector{ResultPathInfoStruct}}}

Take individual timestep and calculate its corresponding Yt_hat.

...

Arguments

gold_ind::Union{Nothing, Vector}=nothing: for in goldpathinfo mode

Shat_val::Union{Nothing, Matrix}=nothing: for goldpathinfo mode

check_gold_path::Bool=false: if true, turn on goldpathinfo mode

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum candidates to keep in the results

threshold::AbstractFloat=0.1:the value set for the support such that if the support of a n-gram is higher than this value, select the n-gram anyway

is_tolerant::Bool=false: if true, select a specified number of n-grams whose supports are below threshold and above tolerance to be added to the path

tolerance::AbstractFloat=(-1000.0): the value set in tolerant mode such that if the support for a n-gram is inbetween this value and the threshold and the max_tolerance number has not been reached, then allow this n-gram to be added to the path

max_tolerance::Int64=4: maximum number of nodes allowed in a path

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: if true, the dataset target is assumed to be tokenized

sep_token::Union{Nothing, String, Char}=nothing: separate token

keep_sep::Bool=false:if true, keep separators in results as well

target_col::Union{String, :Symbol}=:Words: target column names

issparse::Symbol=:auto: Mt matrix output format mode

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information will be printed out

Examples

res_train, gpi_train = JudiLing.learn_paths(
  latin_train,
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  gold_ind=cue_obj_train.gold_ind,
  Shat_val=Shat_train,
  check_gold_path=true,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  sep_token="_",
  keep_sep=false,
  target_col=:Word,
  issparse=:dense,
  verbose=false)

res_val, gpi_val = JudiLing.learn_paths(
  latin_train,
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  gold_ind=cue_obj_val.gold_ind,
  Shat_val=Shat_val,
  check_gold_path=true,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  is_tolerant=true,
  tolerance=0.1,
  max_tolerance=0,
  tokenized=false,
  sep_token="-",
  keep_sep=false,
  target_col=:Word,
  issparse=:dense,
  verbose=false)

...

JudiLing.build_paths — Method

buildpaths(::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict,::Array) -> ::Vector{Vector{ResultPathInfoStruct}}

build_paths function is a shortcut algorithm for finding paths that only takes the n-grams of that which are close to the targets.

...

Arguments

rC::Union{Nothing, Matrix}=nothing: correlation Matrix of C and Chat, pass it to save computing time

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum candidates to keep in the results

n_neighbors::Int64=10: find indices only in top n neighbors

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: if true, the dataset target is assumed to be tokenized

sep_token::Union{Nothing, String, Char}=nothing: separate token

target_col::Union{String, :Symbol}=:Words: target column names

verbose::Bool=false: if true, more information will be printed out

Examples

JudiLing.build_paths(
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

JudiLing.build_paths(
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

...

JudiLing.eval_can — Method

evalcan(::Vector{Vector{Tuple{Vector{Int64}, Int64}}},::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Dict,::Int64,::Bool) -> ::Array{Array{ResultPathInfoStruct,1},1}

Evaluate each candidate with regard to its predicted semantic vector.

JudiLing.find_top_feature_indices — Method

findtopfeature_indices(::Matrix, ::Array) -> ::Vector{Vector{Int64}}

Find all indices of the top n closest neighbors for a given target.