Find Paths

JudiLing.Result_Path_Info_Struct — Type

Store paths' information built by learn_paths or build_paths

JudiLing.Gold_Path_Info_Struct — Type

Store gold paths' information including indices and indices' support and total support. It can be used to evaluate how low the threshold needs to be set in order to find most of the correct paths or if set very low, all of the correct paths.

JudiLing.learn_paths — Function

learn_paths(::DataFrame, ::DataFrame, ::SparseMatrixCSC, ::Union{SparseMatrixCSC, Matrix}, ::Union{SparseMatrixCSC, Matrix}, ::Matrix, ::SparseMatrixCSC, ::Dict) -> ::Union{Tuple{Vector{Vector{Result_Path_Info_Struct}}, Vector{Gold_Path_Info_Struct}}, Vector{Vector{Result_Path_Info_Struct}}}

A sequence finding algorithm using discrimination learning to predict, for a given word, which n-grams are best supported for a given position in the sequence of n-grams.

...

Obligatory Arguments

data::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

C_train::SparseMatrixCSC: the C matrix for training dataset

S_val::Union{SparseMatrixCSC, Matrix}: the S matrix for validation dataset

F_train::Union{SparseMatrixCSC, Matrix}: the F matrix for training dataset

Chat_val::Matrix: the Chat matrix for validation dataset

A::SparseMatrixCSC: the adjacency matrix

i2f::Dict: the dictionary returning features given indices

f2i::Dict: the dictionary returning indices given features

Optional Arguments

gold_ind::Union{Nothing, Vector}=nothing: gold paths' indices

Shat_val::Union{Nothing, Matrix}=nothing: the Shat matrix for the validation dataset

check_gold_path::Bool=false: if true, return a list of support values for the gold path; this information is returned as second output value

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum number of candidates to consider

threshold::Float64=0.1:the value set for the support such that if the support of an n-gram is higher than this value, the n-gram will be taking into consideration

is_tolerant::Bool=false: if true, select a specified number (given by max_tolerance) of n-grams whose supports are below threshold but above a second tolerance threshold to be added to the path

tolerance::Float64=(-1000.0): the value set for the second threshold (in tolerant mode) such that if the support for an n-gram is in between this value and the threshold and the max_tolerance number has not been reached, then allow this n-gram to be added to the path

max_tolerance::Int64=4: maximum number of n-grams allowed in a path

grams::Int64=3: the number n of grams that make up an n-gram

tokenized::Bool=false: if true, the dataset target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator token

keep_sep::Bool=false:if true, keep separators in cues

target_col::Union{String, :Symbol}=:Words: the column name for target strings

issparse::Symbol=:auto: control of whether output of Mt matrix is a dense matrix or a sparse matrix

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information is printed

Examples

# basic usage without tokenization
res = JudiLing.learn_paths(
  latin,
  latin,
  cue_obj.C,
  S,
  F,
  Chat,
  A,
  cue_obj.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  keep_sep=false,
  target_col=:Word,
  verbose=true)

# basic usage with tokenization
res = JudiLing.learn_paths(
  french,
  french,
  cue_obj.C,
  S,
  F,
  Chat,
  A,
  cue_obj.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=true,
  sep_token="-",
  keep_sep=true,
  target_col=:Syllables,
  verbose=true)

# basic usage for validation data
res_val = JudiLing.learn_paths(
  latin_train,
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  keep_sep=false,
  target_col=:Word,
  verbose=true)

# turn on tolerance mode
res_val = JudiLing.learn_paths(
  ...
  threshold=0.1,
  is_tolerant=true,
  tolerance=-0.1,
  max_tolerance=4,
  ...)

# turn on check gold paths mode
res_train, gpi_train = JudiLing.learn_paths(
  ...
  gold_ind=cue_obj_train.gold_ind,
  Shat_val=Shat_train,
  check_gold_path=true,
  ...)

res_val, gpi_val = JudiLing.learn_paths(
  ...
  gold_ind=cue_obj_val.gold_ind,
  Shat_val=Shat_val,
  check_gold_path=true,
  ...)

# control over sparsity
res_val = JudiLing.learn_paths(
  ...
  issparse=:auto,
  sparse_ratio=0.2,
  ...)

...

JudiLing.build_paths — Function

build_paths(::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict,::Array) -> ::Vector{Vector{Result_Path_Info_Struct}}

the build_paths function constructs paths by only considering those n-grams that are close to the target. It first takes the predicted c-hat vector and finds the closest n neighbors in the C matrix. Then it selects all n-grams of these neighbors, and constructs all valid paths with those n-grams. The path producing the best correlation with the target semantic vector (through synthesis by analysis) is selected.

...

Obligatory Arguments

data::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

C_train::SparseMatrixCSC: the C matrix for the training dataset

S_val::Union{SparseMatrixCSC, Matrix}: the S matrix for the validation dataset

F_train::Union{SparseMatrixCSC, Matrix}: the F matrix for the training dataset

Chat_val::Matrix: the Chat matrix for the validation dataset

A::SparseMatrixCSC: the adjacency matrix

i2f::Dict: the dictionary returning features given indices

C_train_ind::Array: the gold paths' indices for the training dataset

Optional Arguments

rC::Union{Nothing, Matrix}=nothing: correlation Matrix of C and Chat, specify to save computing time

max_t::Int64=15: maximum number of timesteps

max_can::Int64=10: maximum number of candidates to consider

n_neighbors::Int64=10: the top n form neighbors to be considered

grams::Int64=3: the number n of grams that make up n-grams

tokenized::Bool=false: if true, the dataset target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator

target_col::Union{String, :Symbol}=:Words: the column name for target strings

verbose::Bool=false: if true, more information will be printed

Examples

# training dataset
JudiLing.build_paths(
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

# validation dataset
JudiLing.build_paths(
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

...

JudiLing.learn_paths — Method

learn_paths(::DataFrame, ::DataFrame, ::SparseMatrixCSC, ::Union{SparseMatrixCSC, Matrix}, ::Union{SparseMatrixCSC, Matrix}, ::Matrix, ::SparseMatrixCSC, ::Dict) -> ::Union{Tuple{Vector{Vector{Result_Path_Info_Struct}}, Vector{Gold_Path_Info_Struct}}, Vector{Vector{Result_Path_Info_Struct}}}

A sequence finding algorithm using discrimination learning to predict, for a given word, which n-grams are best supported for a given position in the sequence of n-grams.

...

Obligatory Arguments

data::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

C_train::SparseMatrixCSC: the C matrix for training dataset

S_val::Union{SparseMatrixCSC, Matrix}: the S matrix for validation dataset

F_train::Union{SparseMatrixCSC, Matrix}: the F matrix for training dataset

Chat_val::Matrix: the Chat matrix for validation dataset

A::SparseMatrixCSC: the adjacency matrix

i2f::Dict: the dictionary returning features given indices

f2i::Dict: the dictionary returning indices given features

Optional Arguments

gold_ind::Union{Nothing, Vector}=nothing: gold paths' indices

Shat_val::Union{Nothing, Matrix}=nothing: the Shat matrix for the validation dataset

check_gold_path::Bool=false: if true, return a list of support values for the gold path; this information is returned as second output value

max_t::Int64=15: maximum timestep

max_can::Int64=10: maximum number of candidates to consider

threshold::Float64=0.1:the value set for the support such that if the support of an n-gram is higher than this value, the n-gram will be taking into consideration

is_tolerant::Bool=false: if true, select a specified number (given by max_tolerance) of n-grams whose supports are below threshold but above a second tolerance threshold to be added to the path

tolerance::Float64=(-1000.0): the value set for the second threshold (in tolerant mode) such that if the support for an n-gram is in between this value and the threshold and the max_tolerance number has not been reached, then allow this n-gram to be added to the path

max_tolerance::Int64=4: maximum number of n-grams allowed in a path

grams::Int64=3: the number n of grams that make up an n-gram

tokenized::Bool=false: if true, the dataset target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator token

keep_sep::Bool=false:if true, keep separators in cues

target_col::Union{String, :Symbol}=:Words: the column name for target strings

issparse::Symbol=:auto: control of whether output of Mt matrix is a dense matrix or a sparse matrix

sparse_ratio::Float64=0.2: the ratio to decide whether a matrix is sparse

verbose::Bool=false: if true, more information is printed

Examples

# basic usage without tokenization
res = JudiLing.learn_paths(
  latin,
  latin,
  cue_obj.C,
  S,
  F,
  Chat,
  A,
  cue_obj.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  keep_sep=false,
  target_col=:Word,
  verbose=true)

# basic usage with tokenization
res = JudiLing.learn_paths(
  french,
  french,
  cue_obj.C,
  S,
  F,
  Chat,
  A,
  cue_obj.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=true,
  sep_token="-",
  keep_sep=true,
  target_col=:Syllables,
  verbose=true)

# basic usage for validation data
res_val = JudiLing.learn_paths(
  latin_train,
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  max_t=max_t,
  max_can=10,
  grams=3,
  threshold=0.1,
  tokenized=false,
  keep_sep=false,
  target_col=:Word,
  verbose=true)

# turn on tolerance mode
res_val = JudiLing.learn_paths(
  ...
  threshold=0.1,
  is_tolerant=true,
  tolerance=-0.1,
  max_tolerance=4,
  ...)

# turn on check gold paths mode
res_train, gpi_train = JudiLing.learn_paths(
  ...
  gold_ind=cue_obj_train.gold_ind,
  Shat_val=Shat_train,
  check_gold_path=true,
  ...)

res_val, gpi_val = JudiLing.learn_paths(
  ...
  gold_ind=cue_obj_val.gold_ind,
  Shat_val=Shat_val,
  check_gold_path=true,
  ...)

# control over sparsity
res_val = JudiLing.learn_paths(
  ...
  issparse=:auto,
  sparse_ratio=0.2,
  ...)

...

JudiLing.build_paths — Method

build_paths(::DataFrame,::SparseMatrixCSC,::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Matrix,::SparseMatrixCSC,::Dict,::Array) -> ::Vector{Vector{Result_Path_Info_Struct}}

the build_paths function constructs paths by only considering those n-grams that are close to the target. It first takes the predicted c-hat vector and finds the closest n neighbors in the C matrix. Then it selects all n-grams of these neighbors, and constructs all valid paths with those n-grams. The path producing the best correlation with the target semantic vector (through synthesis by analysis) is selected.

...

Obligatory Arguments

data::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

C_train::SparseMatrixCSC: the C matrix for the training dataset

S_val::Union{SparseMatrixCSC, Matrix}: the S matrix for the validation dataset

F_train::Union{SparseMatrixCSC, Matrix}: the F matrix for the training dataset

Chat_val::Matrix: the Chat matrix for the validation dataset

A::SparseMatrixCSC: the adjacency matrix

i2f::Dict: the dictionary returning features given indices

C_train_ind::Array: the gold paths' indices for the training dataset

Optional Arguments

rC::Union{Nothing, Matrix}=nothing: correlation Matrix of C and Chat, specify to save computing time

max_t::Int64=15: maximum number of timesteps

max_can::Int64=10: maximum number of candidates to consider

n_neighbors::Int64=10: the top n form neighbors to be considered

grams::Int64=3: the number n of grams that make up n-grams

tokenized::Bool=false: if true, the dataset target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator

target_col::Union{String, :Symbol}=:Words: the column name for target strings

verbose::Bool=false: if true, more information will be printed

Examples

# training dataset
JudiLing.build_paths(
  latin_train,
  cue_obj_train.C,
  S_train,
  F_train,
  Chat_train,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

# validation dataset
JudiLing.build_paths(
  latin_val,
  cue_obj_train.C,
  S_val,
  F_train,
  Chat_val,
  A,
  cue_obj_train.i2f,
  cue_obj_train.gold_ind,
  max_t=max_t,
  n_neighbors=10,
  verbose=false
  )

...

JudiLing.eval_can — Method

eval_can(::Vector{Vector{Tuple{Vector{Int64}, Int64}}},::Union{SparseMatrixCSC, Matrix},::Union{SparseMatrixCSC, Matrix},::Dict,::Int64,::Bool) -> ::Array{Array{Result_Path_Info_Struct,1},1}

Calculate for each candidate path the correlation between predicted semantic vector and the gold standard semantic vector, and select as target for production the path with the highest correlation.

JudiLing.find_top_feature_indices — Method

find_top_feature_indices(::Matrix, ::Array) -> ::Vector{Vector{Int64}}

Find all indices for the n-grams of the top n closest neighbors of a given target.