Make Semantic Matrix

JudiLing.PS_Matrix_Struct — Type

A structure that stores the discrete semantic vectors: pS is the discrete semantic matrix; f2i is a dictionary returning the indices for features; i2f is a dictionary returning the features for indices.

JudiLing.make_pS_matrix — Function

Make discrete semantic matrix.

JudiLing.make_S_matrix — Function

Make simulated semantic matrix.

JudiLing.make_pS_matrix — Method

make_pS_matrix(::DataFrame) -> ::PS_Matrix_Struct

Create a discrete semantic matrix given a dataframe.

...

Obligatory Arguments

data::DataFrame: the dataset

Optional Arguments

features_col::Symbol=:CommunicativeIntention: the column name for target

sep_token::String="_": separator

Examples

s_obj_train = JudiLing.make_pS_matrix(
  utterance,
  features_col=:CommunicativeIntention,
  sep_token="_")

...

JudiLing.make_pS_matrix — Method

make_pS_matrix(::DataFrame, ::PS_Matrix_Struct) -> ::PS_Matrix_Struct

Construct discrete semantic matrix for the validation datasets given by the exemplar in the dataframe, and given the S matrix for the training datasets.

...

Obligatory Arguments

utterances::DataFrame: the dataset

utterances_train::PS_Matrix_Struct: training PS object

Optional Arguments

features_col::Symbol=:CommunicativeIntention: the column name for target

sep_token::String="_": separator

Examples

s_obj_val = JudiLing.make_pS_matrix(
  utterance_val,
  s_obj_train,
  features_col=:CommunicativeIntention,
  sep_token="_")

...

JudiLing.make_S_matrix — Method

make_S_matrix(::DataFrame, ::Vector, ::Vector) -> ::Matrix

Create simulated semantic matrix for the training datasets, given the input data of a vector specified contex lexemes and a vector specified gramatic lexemes. The semantic vector of a word form is constructed summing semantic vectors of content and gramatic lexemes.

...

Obligatory Arguments

data::DataFrame: the dataset

base::Vector: context lexemes

inflections::Vector: grammatic lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train = JudiLing.make_S_matrix(
  french,
  ["Lexeme"],
  ["Tense","Aspect","Person","Number","Gender","Class","Mood"],
  ncol=200)

# deep mode
S_train = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  isdeep=true,
  ...)

# non-deep mode
S_train = JudiLing.make_S_matrix(
  ...
  isdeep=false,
  ...)

# add additional Gaussian noise
S_train = JudiLing.make_S_matrix(
  ...
  add_noise=true,
  sd_noise=1,
  ...)

# further control of means and standard deviations
S_train = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  sd_base=4,
  sd_inflection=4,
  sd_noise=1,
  ...)

...

JudiLing.make_S_matrix — Method

make_S_matrix(::DataFrame, ::DataFrame, ::Vector, ::Vector) -> ::Tuple{Matrix, Matrix}

Create simulated semantic matrix for the validation datasets, given the input data of a vector specified contex lexemes and a vector specified gramatic lexemes. The semantic vector of a word form is constructed summing semantic vectors of content and gramatic lexemes.

...

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

inflections::Vector: grammatic lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_S_matrix(
  french,
  french_val,
  ["Lexeme"],
  ["Tense","Aspect","Person","Number","Gender","Class","Mood"],
  ncol=200)

# deep mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  isdeep=true,
  ...)

# non-deep mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  isdeep=false,
  ...)

# add additional Gaussian noise
S_train, S_val = JudiLing.make_S_matrix(
  ...
  add_noise=true,
  sd_noise=1,
  ...)

# further control of means and standard deviations
S_train, S_val = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  sd_base=4,
  sd_inflection=4,
  sd_noise=1,
  ...)

...

JudiLing.make_S_matrix — Method

make_S_matrix(::DataFrame, ::Vector) -> ::Matrix

Create simulated semantic matrix for the training datasets with only base features, given the input data of a vector specified contex lexemes and a vector specified gramatic lexemes. The semantic vector of a word form is constructed summing semantic vectors of content and gramatic lexemes.

...

Obligatory Arguments

data::DataFrame: the dataset

base::Vector: context lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_base::Int64=4: the sd of base features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train = JudiLing.make_S_matrix(
  french,
  ["Lexeme"],
  ncol=200)

# deep mode
S_train = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  isdeep=true,
  ...)

# non-deep mode
S_train = JudiLing.make_S_matrix(
  ...
  isdeep=false,
  ...)

# add additional Gaussian noise
S_train = JudiLing.make_S_matrix(
  ...
  add_noise=true,
  sd_noise=1,
  ...)

# further control of means and standard deviations
S_train = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  sd_base=4,
  sd_inflection=4,
  sd_noise=1,
  ...)

...

JudiLing.make_S_matrix — Method

make_S_matrix(::DataFrame, ::DataFrame, ::Vector) -> ::Tuple{Matrix, Matrix}

Create simulated semantic matrix for the validation datasets with only base features, given the input data of a vector specified contex lexemes and a vector specified gramatic lexemes. The semantic vector of a word form is constructed summing semantic vectors of content and gramatic lexemes.

...

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_base::Int64=4: the sd of base features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_S_matrix(
  french,
  french_val,
  ["Lexeme"],
  ncol=200)

# deep mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  isdeep=true,
  ...)

# non-deep mode
S_train, S_val = JudiLing.make_S_matrix(
  ...
  isdeep=false,
  ...)

# add additional Gaussian noise
S_train, S_val = JudiLing.make_S_matrix(
  ...
  add_noise=true,
  sd_noise=1,
  ...)

# further control of means and standard deviations
S_train, S_val = JudiLing.make_S_matrix(
  ...
  sd_base_mean=1,
  sd_inflection_mean=1,
  sd_base=4,
  sd_inflection=4,
  sd_noise=1,
  ...)

...