Make Semantic Matrix

JudiLing.PS_Matrix_Struct — Type

This a struct that store all information about prelinguistic and their feature indices. pS is the cue matrix f2i is the dictionary return indices giving features i2f is in another hand return features when giving indices

JudiLing.make_pS_matrix — Function

This is the function that make prelinguistic semantic matrix.

JudiLing.make_S_matrix — Function

This is the function that make simulated semantic matrix.

JudiLing.make_pS_matrix — Method

makepSmatrix(::DataFrame)

This is a function that create prelinguistic matrix giving a csv file.

...

Arguments

features_col::Symbol=:CommunicativeIntention: the column name for communicative intention

sep_token::String="_": the seperated token in the communicative intention column

Examples

utterance = CSV.DataFrame!(CSV.File(joinpath("data", "utterance_mini.csv")))
s_obj_train = JudiLing.make_pS_matrix(utterance)

...

JudiLing.make_pS_matrix — Method

makepSmatrix(::DataFrame, PSMatrixStruct)

This is a function that construct prelinguistic matrix giving utterances and training sobj. The feature indices should maintain the same as thoes in sobj.

...

Arguments

features_col::Symbol=:CommunicativeIntention: the column name for communicative intention

sep_token::String="_": the seperated token in the communicative intention column

Examples

utterance = CSV.DataFrame!(CSV.File(joinpath("data", "utterance_mini.csv")))
s_obj_train = JudiLing.make_pS_matrix(utterance)
s_obj_val = JudiLing.make_pS_matrix(utterance_val, s_obj_train)

...

JudiLing.make_S_matrix — Method

makeSmatrix(::DataFrame, ::Vector, ::Vector)

This is a function that create simulated semantic matrix. Give each feature a random semantic vector, and sum up all features to compose the semantic vector.

...

Arguments

ncol::Int64=200: the dimension size of vectors, usually the same as cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if in deep mode, mean of each feature is also randomized

add_noise::Bool=true: whether to add noise at the end of construction

sd_noise::Int64=1: the sd of the noise matrix

normalized::Bool=false: if true, most of the values range between 1 and -1, it may exceeds 1 or -1 depending on the sd

Examples

# Examples
french = CSV.DataFrame!(CSV.File(joinpath("data", "french_mini.csv")))
S_train = JudiLing.make_S_matrix(
  french,
  ["Lexeme"],
  ["Tense","Aspect","Person","Number","Gender","Class","Mood"])

...

JudiLing.make_S_matrix — Method

makeSmatrix(::DataFrame, ::DataFrame, ::Vector, ::Vector)

This is a function that create validation simulated semantic matrix. Give each feature a random semantic vector, and sum up all features to compose the semantic vector.

...

Arguments

ncol::Int64=200: the dimension size of vectors, usually the same as cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if in deep mode, mean of each feature is also randomized

add_noise::Bool=true: whether to add noise at the end of construction

sd_noise::Int64=1: the sd of the noise matrix

normalized::Bool=false: if normalized, values of matrix maintain close between 1 and -1

Examples

# Examples
french = CSV.DataFrame!(CSV.File(joinpath("data", "french_mini.csv")))
french_val = french[100:end,:]
S_train, S_val = JudiLing.make_S_matrix(
    french,
    french_val,
    ["Lexeme"],
    ["Tense","Aspect","Person","Number","Gender","Class","Mood"])

...

JudiLing.make_S_matrix — Method

makeSmatrix(::DataFrame, ::Vector)

This is a function that create simulated semantic matrix, provided for dataset that only have base features. Give each feature a random semantic vector, and sum up all features to compose the semantic vector.

...

Arguments

ncol::Int64=200: the dimension size of vectors, usually the same as cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_base::Int64=4: the sd of base features

seed::Int64=314: the random seed

isdeep::Bool=true: if in deep mode, mean of each feature is also randomized

add_noise::Bool=true: whether to add noise at the end of construction

sd_noise::Int64=1: the sd of the noise matrix

Examples

french = CSV.DataFrame!(CSV.File(joinpath("data", "french_mini.csv")))

S_train = JudiLing.make_S_matrix(
  french,
  base=["Lexeme"])=

...

JudiLing.make_S_matrix — Method

makeSmatrix(::DataFrame, ::DataFrame, ::Vector)

This is a function that create validation simulated semantic matrix, provided for dataset that only have base features. Give each feature a random semantic vector, and sum up all features to compose the semantic vector.

...

Arguments

ncol::Int64=200: the dimension size of vectors, usually the same as cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_base::Int64=4: the sd of base features

seed::Int64=314: the random seed

isdeep::Bool=true: if in deep mode, mean of each feature is also randomized

add_noise::Bool=true: whether to add noise at the end of construction

sd_noise::Int64=1: the sd of the noise matrix

Examples

french = CSV.DataFrame!(CSV.File(joinpath("data", "french_mini.csv")))
french_val = french[100:end,:]
S_train, S_val = JudiLing.make_S_matrix(
    french,
    french_val,
    base=["Lexeme"])

...