Make Semantic Matrix

JudiLing.PS_Matrix_Struct — Type

A structure that stores the discrete semantic vectors: pS is the discrete semantic matrix; f2i is a dictionary returning the indices for features; i2f is a dictionary returning the features for indices.

JudiLing.L_Matrix_Struct — Type

A structure that stores Lexome semantic vectors: L is Lexome semantic matrix; f2i is a dictionary returning the indices for features; i2f is a dictionary returning the features for indices.

JudiLing.make_pS_matrix — Function

Make discrete semantic matrix.

JudiLing.make_S_matrix — Function

Make simulated semantic matrix.

JudiLing.make_L_matrix — Function

Make simulated lexome matrix.

JudiLing.make_combined_S_matrix — Function

Make combined simulated S matrices, where combined features from both training datasets and validation datasets

JudiLing.make_combined_L_matrix — Function

Make combined simulated Lexome matrix, where combined features from both training datasets and validation datasets

JudiLing.make_pS_matrix — Method

make_pS_matrix(utterances)

Create a discrete semantic matrix given a dataframe.

Obligatory Arguments

data::DataFrame: the dataset

Optional Arguments

features_col::Symbol=:CommunicativeIntention: the column name for target

sep_token::String="_": separator

Examples

s_obj_train = JudiLing.make_pS_matrix(
    utterance,
    features_col=:CommunicativeIntention,
    sep_token="_")

JudiLing.make_pS_matrix — Method

make_pS_matrix(utterances, utterances_train)

Construct discrete semantic matrix for the validation datasets given by the exemplar in the dataframe, and given the S matrix for the training datasets.

Obligatory Arguments

utterances::DataFrame: the dataset

utterances_train::PS_Matrix_Struct: training PS object

Optional Arguments

features_col::Symbol=:CommunicativeIntention: the column name for target

sep_token::String="_": separator

Examples

s_obj_val = JudiLing.make_pS_matrix(
    utterance_val,
    s_obj_train,
    features_col=:CommunicativeIntention,
    sep_token="_")

JudiLing.make_S_matrix — Method

make_S_matrix(data::DataFrame, base::Vector, inflections::Vector)

Create simulated semantic matrix for the training datasets, given the input data of a vector specified contex lexemes and a vector specified gramatic lexemes. The semantic vector of a word form is constructed summing semantic vectors of content and gramatic lexemes.

Obligatory Arguments

data::DataFrame: the dataset

base::Vector: context lexemes

inflections::Vector: grammatic lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train = JudiLing.make_S_matrix(
    french,
    ["Lexeme"],
    ["Tense","Aspect","Person","Number","Gender","Class","Mood"],
    ncol=200)

# deep mode
S_train = JudiLing.make_S_matrix(
    ...
    sd_base_mean=1,
    sd_inflection_mean=1,
    isdeep=true,
    ...)

# non-deep mode
S_train = JudiLing.make_S_matrix(
    ...
    isdeep=false,
    ...)

# add additional Gaussian noise
S_train = JudiLing.make_S_matrix(
    ...
    add_noise=true,
    sd_noise=1,
    ...)

# further control of means and standard deviations
S_train = JudiLing.make_S_matrix(
    ...
    sd_base_mean=1,
    sd_inflection_mean=1,
    sd_base=4,
    sd_inflection=4,
    sd_noise=1,
    ...)

JudiLing.make_S_matrix — Method

make_S_matrix(data_train::DataFrame, data_val::DataFrame, base::Vector, inflections::Vector)

Create simulated semantic matrix for the validation datasets, given the input data of a vector specified contex lexemes and a vector specified gramatic lexemes. The semantic vector of a word form is constructed summing semantic vectors of content and gramatic lexemes.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

inflections::Vector: grammatic lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_S_matrix(
    french,
    french_val,
    ["Lexeme"],
    ["Tense","Aspect","Person","Number","Gender","Class","Mood"],
    ncol=200)

# deep mode
S_train, S_val = JudiLing.make_S_matrix(
    ...
    sd_base_mean=1,
    sd_inflection_mean=1,
    isdeep=true,
  ...)

# non-deep mode
S_train, S_val = JudiLing.make_S_matrix(
    ...
    isdeep=false,
    ...)

# add additional Gaussian noise
S_train, S_val = JudiLing.make_S_matrix(
    ...
    add_noise=true,
    sd_noise=1,
    ...)

# further control of means and standard deviations
S_train, S_val = JudiLing.make_S_matrix(
    ...
    sd_base_mean=1,
    sd_inflection_mean=1,
    sd_base=4,
    sd_inflection=4,
    sd_noise=1,
    ...)

JudiLing.make_S_matrix — Method

make_S_matrix(data::DataFrame, base::Vector)

Create simulated semantic matrix for the training datasets with only base features, given the input data of a vector specified contex lexemes and a vector specified gramatic lexemes. The semantic vector of a word form is constructed summing semantic vectors of content and gramatic lexemes.

Obligatory Arguments

data::DataFrame: the dataset

base::Vector: context lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_base::Int64=4: the sd of base features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train = JudiLing.make_S_matrix(
    french,
    ["Lexeme"],
    ncol=200)

# deep mode
S_train = JudiLing.make_S_matrix(
    ...
    sd_base_mean=1,
    sd_inflection_mean=1,
    isdeep=true,
    ...)

# non-deep mode
S_train = JudiLing.make_S_matrix(
    ...
    isdeep=false,
    ...)

# add additional Gaussian noise
S_train = JudiLing.make_S_matrix(
    ...
    add_noise=true,
    sd_noise=1,
    ...)

# further control of means and standard deviations
S_train = JudiLing.make_S_matrix(
    ...
    sd_base_mean=1,
    sd_inflection_mean=1,
    sd_base=4,
    sd_inflection=4,
    sd_noise=1,
    ...)

JudiLing.make_S_matrix — Method

make_S_matrix(data_train::DataFrame, data_val::DataFrame, base::Vector)

Create simulated semantic matrix for the validation datasets with only base features, given the input data of a vector specified contex lexemes and a vector specified gramatic lexemes. The semantic vector of a word form is constructed summing semantic vectors of content and gramatic lexemes.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_base::Int64=4: the sd of base features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_S_matrix(
    french,
    french_val,
    ["Lexeme"],
    ncol=200)

# deep mode
S_train, S_val = JudiLing.make_S_matrix(
    ...
    sd_base_mean=1,
    sd_inflection_mean=1,
    isdeep=true,
    ...)

# non-deep mode
S_train, S_val = JudiLing.make_S_matrix(
    ...
    isdeep=false,
    ...)

# add additional Gaussian noise
S_train, S_val = JudiLing.make_S_matrix(
    ...
    add_noise=true,
    sd_noise=1,
    ...)

# further control of means and standard deviations
S_train, S_val = JudiLing.make_S_matrix(
    ...
    sd_base_mean=1,
    sd_inflection_mean=1,
    sd_base=4,
    sd_inflection=4,
    sd_noise=1,
    ...)

JudiLing.make_S_matrix — Method

make_S_matrix(data_train::DataFrame, data_val::DataFrame, pyndl_weights::Pyndl_Weight_Struct, n_features_columns::Vector)

Create semantic matrix for pyndl mode

JudiLing.make_S_matrix — Method

make_S_matrix(data_train::DataFrame, base::Vector, inflections::Vector, L::L_Matrix_Struct)

Create simulated semantic matrix where lexome matrix is available.

Obligatory Arguments

data::DataFrame: the dataset

base::Vector: context lexemes

inflections::Vector: grammatic lexemes

L::L_Matrix_Struct: the lexome matrix

Optional Arguments

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S1 = JudiLing.make_S_matrix(
    latin,
    ["Lexeme"],
    ["Person","Number","Tense","Voice","Mood"],
     L1,
     add_noise=true,
    sd_noise=1,
    normalized=false
    )

JudiLing.make_S_matrix — Method

make_S_matrix(data_train::DataFrame, data_val::Union{DataFrame, Nothing}, base::Vector, L::L_Matrix_Struct)

Create simulated semantic matrix where lexome matrix is available.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

L::L_Matrix_Struct: the lexome matrix

Optional Arguments

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S1, S2 = JudiLing.make_S_matrix(
     latin,
    latin_val,
    ["Lexeme"],
    L1,
    add_noise=true,
    sd_noise=1,
    normalized=false
    )

JudiLing.make_S_matrix — Method

make_S_matrix(data::DataFrame, base::Vector, L::L_Matrix_Struct)

Create simulated semantic matrix where lexome matrix is available.

Obligatory Arguments

data::DataFrame: the dataset

base::Vector: context lexemes

L::L_Matrix_Struct: the lexome matrix

Optional Arguments

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S1 = JudiLing.make_S_matrix(
    latin,
    ["Lexeme"],
    L1,
    add_noise=true,
    sd_noise=1,
    normalized=false
    )

JudiLing.make_S_matrix — Method

make_S_matrix(data_train::DataFrame, data_val::DataFrame, base::Vector, inflections::Vector, L::L_Matrix_Struct)

Create simulated semantic matrix where lexome matrix is available.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

inflections::Vector: grammatic lexemes

L::L_Matrix_Struct: the lexome matrix

Optional Arguments

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S1, S2 = JudiLing.make_S_matrix(
    latin,
    latin_val,
    ["Lexeme"],
    ["Person","Number","Tense","Voice","Mood"],
    L1,
    add_noise=true,
    sd_noise=1,
    normalized=false
    )

JudiLing.make_L_matrix — Method

make_L_matrix(data::DataFrame, base::Vector)

Create Lexome Matrix with simulated semantic vectors where there are only base features.

Obligatory Arguments

data::DataFrame: the dataset

base::Vector: context lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_base::Int64=4: the sd of base features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

Examples

# basic usage
L = JudiLing.make_L_matrix(
    latin,
    ["Lexeme"],
    ncol=200)

JudiLing.make_combined_S_matrix — Method

make_combined_S_matrix(data_train::DataFrame, data_val::DataFrame, base::Vector, inflections::Vector, L::L_Matrix_Struct)

Create simulated semantic matrix for the training datasets and validation datasets with existing Lexome matrix, where features are combined from both training datasets and validation datasets.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

inflections::Vector: grammatic lexemes

L::L_Matrix_Struct: the Lexome Matrix

Optional Arguments

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_combined_S_matrix(
    latin_train,
    latin_val,
    ["Lexeme"],
    ["Person","Number","Tense","Voice","Mood"],
    L)

JudiLing.make_combined_S_matrix — Method

make_combined_S_matrix(data_train::DataFrame, data_val::Union{DataFrame, Nothing}, base::Vector, L::L_Matrix_Struct)

Create simulated semantic matrix for the training datasets and validation datasets with existing Lexome matrix, where features are combined from both training datasets and validation datasets.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

L::L_Matrix_Struct: the Lexome Matrix

Optional Arguments

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_combined_S_matrix(
    latin_train,
    latin_val,
    ["Lexeme"],
    ["Person","Number","Tense","Voice","Mood"],
    L)

JudiLing.make_combined_S_matrix — Method

make_combined_S_matrix(  data_train::DataFrame, data_val::DataFrame, base::Vector, inflections::Vector)

Create simulated semantic matrix for the training datasets and validation datasets, where features are combined from both training datasets and validation datasets.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

inflections::Vector: grammatic lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_combined_S_matrix(
    latin_train,
    latin_val,
    ["Lexeme"],
    ["Person","Number","Tense","Voice","Mood"],
    ncol=n_features)

JudiLing.make_combined_S_matrix — Method

make_combined_S_matrix(data_train::DataFrame, data_val::DataFrame, base::Vector)

Create simulated semantic matrix for the training datasets and validation datasets, where features are combined from both training datasets and validation datasets.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

add_noise::Bool=true: if true, add additional Gaussian noise

sd_noise::Int64=1: the sd of the Gaussian noise

normalized::Bool=false: if true, most of the values range between 1 and -1, it may slightly exceed between 1 or -1 depending on the sd

Examples

# basic usage
S_train, S_val = JudiLing.make_combined_S_matrix(
    latin_train,
    latin_val,
    ["Lexeme"],
    ["Person","Number","Tense","Voice","Mood"],
    ncol=n_features)

JudiLing.make_combined_L_matrix — Method

make_combined_L_matrix(data_train::DataFrame, data_val::DataFrame, base::Vector, inflections::Vector)

Create Lexome Matrix with simulated semantic vectors, where features are combined from both training datasets and validation datasets.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

inflections::Vector: grammatic lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

Examples

# basic usage
L = JudiLing.make_combined_L_matrix(
    latin_train,
    latin_val,
    ["Lexeme"],
    ["Person","Number","Tense","Voice","Mood"],
    ncol=n_features)

JudiLing.make_combined_L_matrix — Method

make_combined_L_matrix(data_train::DataFrame, data_val::DataFrame, base::Vector)

Create Lexome Matrix with simulated semantic vectors, where features are combined from both training datasets and validation datasets.

Obligatory Arguments

data_train::DataFrame: the training dataset

data_val::DataFrame: the validation dataset

base::Vector: context lexemes

Optional Arguments

ncol::Int64=200: dimension of semantic vectors, usually the same as that of cue vectors

sd_base_mean::Int64=1: the sd mean of base features

sd_inflection_mean::Int64=1: the sd mean of inflectional features

sd_base::Int64=4: the sd of base features

sd_inflection::Int64=4: the sd of inflectional features

seed::Int64=314: the random seed

isdeep::Bool=true: if true, mean of each feature is also randomized

Examples

# basic usage
L = JudiLing.make_combined_L_matrix(
    latin_train,
    latin_val,
    ["Lexeme"],
    ncol=n_features)

JudiLing.L_Matrix_Struct — Method

L_Matrix_Struct(L, sd_base, sd_base_mean, sd_inflection, sd_inflection_mean, base_f, infl_f, base_f2i, infl_f2i, n_base_f, n_infl_f, ncol)

Construct LMatrixStruct with deep mode.

JudiLing.L_Matrix_Struct — Method

L_Matrix_Struct(L, sd_base, sd_inflection, base_f, infl_f, base_f2i, infl_f2i, n_base_f, n_infl_f, ncol)

Construct LMatrixStruct without deep mode.

JudiLing.process_features — Method

process_features(data, feature_cols)

Collect all features given datasets and feature column names.

JudiLing.comp_f_M! — Method

comp_f_M!(L, sd, sd_mean, n_f, ncol, n_b)

Compose feature Matrix with deep mode.

JudiLing.comp_f_M! — Method

comp_f_M!(L, sd, n_f, ncol, n_b)

Compose feature Matrix without deep mode.

JudiLing.merge_f2i — Method

merge_f2i(base_f2i, infl_f2i, n_base_f, n_infl_f)

Merge base f2i dictionary and inflectional f2i dictionary.

JudiLing.lexome_sum — Method

lexome_sum(L, features)

Sum up semantic vector, given lexome vector.

JudiLing.make_St — Method

make_St(L, n, data, base, inflections)

Make S transpose matrix with inflections.

JudiLing.make_St — Method

make_St(L, n, data, base)

Make S transpose matrix without inflections.

JudiLing.add_St_noise! — Method

add_St_noise!(St, sd_noise)

Add noise.

JudiLing.normalize_St! — Method

normalize_St!(St, n_base, n_infl)

Normalize S transpose with inflections.

JudiLing.normalize_St! — Method

normalize_St!(St, n_base)

Normalize S transpose without inflections.