Output

JudiLing.write2csv — Function

Write results into a csv file. This function takes as input the results from the learn_paths and build_paths functions, including the information on gold paths that is optionally returned as second output result.

JudiLing.write2df — Function

Reformat results into a dataframe. This function takes as input the results from the learn_paths and build_paths functions, including the information on gold paths that is optionally returned as second output result.

JudiLing.save_L_matrix — Function

Save lexome matrix into csv file.

JudiLing.load_L_matrix — Function

Load lexome matrix from csv file.

JudiLing.write2csv — Method

write2csv(::Array{Array{Result_Path_Info_Struct,1},1}, ::DataFrame, ::Cue_Matrix_Struct, ::Cue_Matrix_Struct, ::String) -> ::Nothing

Write results into csv file for the results from learn_paths and build_paths.

...

Obligatory Arguments

res::Array{Array{Result_Path_Info_Struct,1},1}: the results from learn_paths or build_paths

data::DataFrame: the dataset

cue_obj_train::Cue_Matrix_Struct: the cue object for training dataset

cue_obj_val::Cue_Matrix_Struct: the cue object for validation dataset

filename::String: the filename

Optional Arguments

grams::Int64=3: the number n in n-gram cues

tokenized::Bool=false: if true, the dataset target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator

start_end_token::Union{String, Char}="#": start and end token in boundary cues

output_sep_token::Union{String, Char}="": output separator

path_sep_token::Union{String, Char}=":": path separator

target_col::Union{String, Symbol}=:Words: the column name for target strings

root_dir::String=".": dir path for project root dir

output_dir::String=".": output dir inside root dir

Examples

# writing results for training data
JudiLing.write2csv(
  res_train,
  latin_train,
  cue_obj_train,
  cue_obj_train,
  "res_latin_train.csv",
  grams=3,
  tokenized=false,
  sep_token=nothing,
  start_end_token="#",
  output_sep_token="",
  path_sep_token=":",
  target_col=:Word,
  root_dir=".",
  output_dir="test_out")

# writing results for validation data
JudiLing.write2csv(
  res_val,
  latin_val,
  cue_obj_train,
  cue_obj_val,
  "res_latin_val.csv",
  grams=3,
  tokenized=false,
  sep_token=nothing,
  start_end_token="#",
  output_sep_token="",
  path_sep_token=":",
  target_col=:Word,
  root_dir=".",
  output_dir="test_out")

...

JudiLing.write2csv — Method

write2csv(::Vector{Gold_Path_Info_Struct}, ::String) -> ::Nothing

Write results into csv file for the gold paths' information optionally returned by learn_paths and build_paths.

...

Obligatory Arguments

gpi::Vector{Gold_Path_Info_Struct}: the gold paths' information

filename::String: the filename

Optional Arguments

root_dir::String=".": dir path for project root dir

output_dir::String=".": output dir inside root dir

Examples

# write gold standard paths to csv for training data
JudiLing.write2csv(
  gpi_train,
  "gpi_latin_train.csv",
  root_dir=".",
  output_dir="test_out"
  )

# write gold standard paths to csv for validation data
JudiLing.write2csv(
  gpi_val,
  "gpi_latin_val.csv",
  root_dir=".",
  output_dir="test_out"
  )

...

JudiLing.write2df — Method

write2df(::Array{Array{Result_Path_Info_Struct,1},1}, ::DataFrame, ::Cue_Matrix_Struct, ::Cue_Matrix_Struct) -> ::DataFrame

Reformat results into a dataframe for the results form learn_paths and build_paths functions.

...

Obligatory Arguments

data::DataFrame: the dataset

Optional Arguments

grams::Int64=3: the number n in n-gram cues

tokenized::Bool=false: if true, the dataset target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: separator

start_end_token::Union{String, Char}="#": start and end token in boundary cues

output_sep_token::Union{String, Char}="": output separator

path_sep_token::Union{String, Char}=":": path separator

target_col::Union{String, Symbol}=:Words: the column name for target strings

Examples

# writing results for training data
JudiLing.write2df(
  res_train,
  latin_train,
  cue_obj_train,
  cue_obj_train,
  grams=3,
  tokenized=false,
  sep_token=nothing,
  start_end_token="#",
  output_sep_token="",
  path_sep_token=":",
  target_col=:Word)

# writing results for validation data
JudiLing.write2df(
  res_val,
  latin_val,
  cue_obj_train,
  cue_obj_val,
  grams=3,
  tokenized=false,
  sep_token=nothing,
  start_end_token="#",
  output_sep_token="",
  path_sep_token=":",
  target_col=:Word)

...

JudiLing.write2df — Method

write2df(::gpi::Vector{Gold_Path_Info_Struct}) -> ::DataFrame

Write results into a dataframe for the gold paths' information optionally returned by learn_paths and build_paths.

...

Obligatory Arguments

gpi::Vector{Gold_Path_Info_Struct}: the gold paths' information

Examples

# write gold standard paths to df for training data
JudiLing.write2csv(gpi_train)

# write gold standard paths to df for validation data
JudiLing.write2csv(gpi_val)

...

JudiLing.save_L_matrix — Method

save_L_matrix(::L_Matrix_Struct, ::String) -> ::Nothing

Save lexome matrix into csv file.

...

Obligatory Arguments

L::L_Matrix_Struct: the lexome matrix struct

filename::String: the filename/filepath

Examples

JudiLing.save_L_matrix(L, joinpath(@__DIR__, "L.csv"))

...

JudiLing.load_L_matrix — Method

load_L_matrix(::String) -> ::L_Matrix_Struct

Load lexome matrix from csv file.

...

Obligatory Arguments

filename::String: the filename/filepath

Examples

L_load = JudiLing.load_L_matrix(joinpath(@__DIR__, "L.csv"))

...