Output

JudiLing.write2csv — Function

write results into csv file, support for path results and gold path information

JudiLing.write2df — Function

create a datafrate with results

JudiLing.write2csv — Method

write2csv(::Array{Array{ResultPathInfoStruct,1},1}, ::DataFrame, ::CueMatrixStruct, ::CueMatrix_Struct, ::String)

write results into csv file, support for path results

...

Arguments

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: whether the target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: what is the separate token

start_end_token::Union{String, Char}="#": token at beginning and ending token

output_sep_token::Union{String, Char}="": output separate token

path_sep_token::Union{String, Char}=":": token glued path nodes together

target_col::Union{String, Symbol}=:Words: the column name for target

root_dir::String=".": dir path for package root dir

output_dir::String=".": output dir inside root dir

Examples

#after you had results from learn_paths or build_paths
JudiLing.write2csv(
  res_train,
  latin_train,
  cue_obj_train,
  cue_obj_train,
  "res_latin_train.csv",
  grams=3,
  tokenized=false,
  sep_token=nothing,
  start_end_token="#",
  output_sep_token="",
  path_sep_token=":",
  target_col=:Word,
  root_dir=".",
  output_dir="test_out")

JudiLing.write2csv(
  res_val,
  latin_val,
  cue_obj_train,
  cue_obj_val,
  "res_latin_val.csv",
  grams=3,
  tokenized=false,
  sep_token=nothing,
  start_end_token="#",
  output_sep_token="",
  path_sep_token=":",
  target_col=:Word,
  root_dir=".",
  output_dir="test_out")

...

JudiLing.write2csv — Method

write2csv(::Vector{GoldPathInfo_Struct}, ::String)

write gold path information into csv file

...

Arguments

root_dir::String=".": dir path for package root dir

output_dir::String=".": output dir inside root dir

Examples

#after you had results from learn_paths or build_paths
JudiLing.write2csv(
  gpi_train,
  "gpi_latin_train.csv",
  root_dir=".",
  output_dir="test_out"
  )

JudiLing.write2csv(
  gpi_val,
  "gpi_latin_val.csv",
  root_dir=".",
  output_dir="test_out"
  )

...

JudiLing.write2df — Method

write2df(::Array{Array{ResultPathInfoStruct,1},1}, ::DataFrame, ::CueMatrixStruct, ::CueMatrix_Struct, ::String)

write results into dataframe, support for path results

...

Arguments

grams::Int64=3: the number of grams for cues

tokenized::Bool=false: whether the target is tokenized

sep_token::Union{Nothing, String, Char}=nothing: what is the separate token

start_end_token::Union{String, Char}="#": token at beginning and ending token

output_sep_token::Union{String, Char}="": output separate token

path_sep_token::Union{String, Char}=":": token glued path nodes together

target_col::Union{String, Symbol}=:Words: the column name for target

Examples

#after you had results from learn_paths or build_paths
JudiLing.write2df(
  res_train,
  latin_train,
  cue_obj_train,
  cue_obj_train,
  "res_latin_train.csv",
  grams=3,
  tokenized=false,
  sep_token=nothing,
  start_end_token="#",
  output_sep_token="",
  path_sep_token=":",
  target_col=:Word)

JudiLing.write2df(
  res_val,
  latin_val,
  cue_obj_train,
  cue_obj_val,
  "res_latin_val.csv",
  grams=3,
  tokenized=false,
  sep_token=nothing,
  start_end_token="#",
  output_sep_token="",
  path_sep_token=":",
  target_col=:Word)

...