Utils

JudiLing.iscorrect — Function

Check whether the predictions are correct.

JudiLing.display_pred — Function

Display prediction nicely.

JudiLing.translate — Function

Translate indices into words or utterances

JudiLing.translate_path — Function

Append indices together to form a path

JudiLing.is_truly_sparse — Function

Check whether a matrix is truly sparse regardless its format, where M is originally a sparse matrix format.

Check whether a matrix is truly sparse regardless its format, where M is originally a dense matrix format.

JudiLing.isattachable — Function

Check whether a gram can attach to another gram.

Check whether a gram can attach to another gram.

JudiLing.iscomplete — Function

Check whether a path is complete.

JudiLing.isstart — Function

Check whether a gram can start a path.

JudiLing.isnovel — Function

Check whether a predicted path is in training data.

JudiLing.check_used_token — Function

Check whether there are tokens already used in dataset as n-gram components.

JudiLing.cal_max_timestep — Function

Calculate the max timestep given training and validation datasets.

Calculate the max timestep given training datasets only.