Utils

JudiLing.iscorrect — Function

check whether the prediction is correct

JudiLing.display_pred — Function

display prediction nicely

JudiLing.translate — Function

translate indices into words or utterances

JudiLing.translate_path — Function

just append indices together

JudiLing.is_truly_sparse — Function

check whether a matrix is truly sparse regardless its format

check whether a matrix is truly sparse regardless its format

JudiLing.isattachable — Function

check a gram is attach to another gram

check a gram is attach to another gram

JudiLing.iscomplete — Function

check a gram could complete a path

JudiLing.isstart — Function

check a gram could start a path

JudiLing.isnovel — Function

check wheter a path is in training data or nor

JudiLing.check_used_token — Function

check whether there are token used in dataset

JudiLing.cal_max_timestep — Function

calculate max timestep given training and validation datasets

calculate max timestep given training dataset